% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualPlot.R
\name{annualPlot}
\alias{annualPlot}
\title{Download plot with annual and weekly data.}
\usage{
annualPlot(package = "packageRank", from = 2019, to = NULL,
  check.package = TRUE, pro.mode = FALSE, sunday.week = TRUE)
}
\arguments{
\item{package}{Character Package name, "R" (for R application) or NULL (for total CRAN package downloads).}

\item{from}{Numeric or Integer Year. "yyyy".}

\item{to}{Numeric or Integer. Year. "yyyy"; NULL uses current year.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{pro.mode}{Logical.}

\item{sunday.week}{Logical. TRUE: week starts on Sunday. FALSE: week starts on Monday.}
}
\description{
Download plot with annual and weekly data.
}
\note{
Adapted from Vesuvius plot at https://github.com/nrennie/tidytuesday/blob/main/2025/2025-05-13/20250513.R
}
