% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyTriplet.R
\name{anyTriplet}
\alias{anyTriplet}
\title{Detect small downloads triplets (prototype).}
\usage{
anyTriplet(pkg, cran_log)
}
\arguments{
\item{pkg}{Character. Package name.}

\item{cran_log}{Object. Package log.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
