% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivePackages.R
\name{archivePackages}
\alias{archivePackages}
\title{Vector of all packages in CRAN archive.}
\usage{
archivePackages(include.date = FALSE)
}
\arguments{
\item{include.date}{Logical. Retrun data frame with package name and last publication date.}
}
\description{
Vector of all packages in CRAN archive.
}
