% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveZeroDownloads.R
\name{archiveZeroDownloads}
\alias{archiveZeroDownloads}
\title{Archive and zero downloads}
\usage{
archiveZeroDownloads(date = Sys.Date() - 1, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
Archive and zero downloads
}
