% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDate.R
\name{available_log0}
\alias{available_log0}
\title{Compute Effective CRAN Log Date Based on Local and UTC Time (prototype).}
\usage{
available_log0(local.date = "2021-01-25", current.date = Sys.Date(),
  current.time = "09:51", tz = Sys.timezone(), upload.time = "17:00",
  warning.msg = TRUE)
}
\arguments{
\item{local.date}{Character. Date of desired log \code{"yyyy-mm-dd"}.}

\item{current.date}{Character. \code{"yyyy-mm-dd"}.}

\item{current.time}{Character. \code{"hh-mm-ss"} \code{"hh-mm"}.}

\item{tz}{Character. Time zone. See OlsonNames().}

\item{upload.time}{Character. UTC upload time for logs "hh:mm" or "hh:mm:ss".}

\item{warning.msg}{Logical. TRUE uses warning() if the function returns the date of the previous available log.}
}
\description{
Compute Effective CRAN Log Date Based on Local and UTC Time (prototype).
}
