% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availablePackages.R
\name{availablePackages}
\alias{availablePackages}
\title{Get vector of available packages on CRAN.}
\usage{
availablePackages()
}
\description{
Scrapes https://cran.r-project.org/web/packages
}
\note{
More up to date than utils::available.packages().
}
