% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads2.R
\name{bio_mo}
\alias{bio_mo}
\title{Monthly package downloads from Bioconductor (prototype).}
\usage{
bio_mo(year = NULL, month = NULL, end.year = NULL,
  end.month = NULL, pkg = NULL)
}
\arguments{
\item{year}{Numeric.}

\item{month}{Numeric.}

\item{end.year}{Numeric.}

\item{end.month}{Numeric.}

\item{pkg}{Character.}
}
\description{
Monthly package downloads from Bioconductor (prototype).
}
\examples{
bio_mo(year = 2015)
bio_mo(year = 2016, pkg = "clusterProfiler")
}
