% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads2.R
\name{bio_yr}
\alias{bio_yr}
\title{Annual/monthly package downloads from Bioconductor (prototype).}
\usage{
bio_yr(year = NULL, end.year = NULL, pkg = NULL)
}
\arguments{
\item{year}{Numeric.}

\item{end.year}{Numeric.}

\item{pkg}{Character.}
}
\description{
Annual/monthly package downloads from Bioconductor (prototype).
}
\examples{
bio_yr(year = 2015)
bio_yr(pkg = "clusterProfiler")
}
