% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads.R
\name{bioconductor_downloads}
\alias{bioconductor_downloads}
\title{Annual/monthly package downloads from Bioconductor (prototype).}
\usage{
bioconductor_downloads(pkg = "clusterProfiler", obs = "year")
}
\arguments{
\item{pkg}{A character vector.}

\item{obs}{Unit of observation: "month" or "year"}
}
\description{
Annual/monthly package downloads from Bioconductor (prototype).
}
