% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{bioconductorDownloads}
\alias{bioconductorDownloads}
\title{Annual/monthly package downloads from Bioconductor.}
\usage{
bioconductorDownloads(package = NULL, from = NULL, to = NULL,
  when = NULL, unit.observation = "month")
}
\arguments{
\item{package}{Character. Vector of package names.}

\item{from}{Start date as \code{yyyy-mm} or \code{yyyy}.}

\item{to}{End date as \code{yyyy-mm} or \code{yyyy}.}

\item{when}{\code{"last-year"}, or \code{"year-to-date"} or \code{"ytd"}.}

\item{unit.observation}{"year" or "month".}
}
\description{
Annual/monthly package downloads from Bioconductor.
}
\examples{
\dontrun{
# all packages
bioconductorDownloads()

# entire history
bioconductorDownloads(package = "clusterProfiler")

# year-to-date
bioconductorDownloads(package = "clusterProfiler", when = "ytd")
bioconductorDownloads(package = "clusterProfiler", when = "year-to-date")

# last 12 months
bioconductorDownloads(package = "clusterProfiler", when = "last-year")

# from 2015 to current year
bioconductorDownloads(package = "clusterProfiler", from = 2015)

# 2010 through 2015 (yearly)
bioconductorDownloads(package = "clusterProfiler", from = 2010, to = 2015,
  unit.observation = "year")

# selected year (yearly)
bioconductorDownloads(package = "clusterProfiler", from = 2015, to = 2015)

# selected year (monthly)
bioconductorDownloads(package = "clusterProfiler", from = "2015-01", to = "2015-12")

# June 2014 through March 2015
bioconductorDownloads(package = "clusterProfiler", from = "2014-06", to = "2015-03")
}
}
