% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorRank.R
\name{bioconductorRank}
\alias{bioconductorRank}
\title{Package download counts and rank percentiles (cross-sectional).}
\usage{
bioconductorRank(pkg = "clusterProfiler", yr.mo = "2019-06",
  count = "download")
}
\arguments{
\item{pkg}{Character. Vector of package name(s).}

\item{yr.mo}{Character. Date.}

\item{count}{Character. "ip" or "download".}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
