% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blog.data}
\alias{blog.data}
\title{blog post data.}
\format{A list with 11 elements.}
\usage{
blog.data
}
\description{
\describe{
   \item{\code{archive.pkg_ver}}{}
   \item{\code{archive.pkg_ver.filtered}}{}
   \item{\code{cran.pkg_ver}}{}
   \item{\code{cran.pkg_ver.filtered}}{}
   \item{\code{dl.ct}}{}
   \item{\code{dl.ct2}}{}
   \item{\code{pkg.ct}}{}
   \item{\code{pkg.ct2}}{}
   \item{\code{oct.data}}{}
   \item{\code{cholera.data}}{}
   \item{\code{ggplot2.data}}{}
   \item{\code{VR.data}}{}
}
}
\keyword{datasets}
