% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{campaignRLE}
\alias{campaignRLE}
\title{Run Length Encoding of First Letter of Packages Downloaded.}
\usage{
campaignRLE(ip, cran_log, case.sensitive = FALSE, concatenate = TRUE)
}
\arguments{
\item{ip}{Numeric. Nominal IP address.}

\item{cran_log}{Object. Package log entries.}

\item{case.sensitive}{Logical.}

\item{concatenate}{Logical.}
}
\description{
Uses rle().
}
\examples{
\dontrun{
campaignRLE(ip = 24851, cran_log = july01)
}
}
