% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{campaigns}
\alias{campaigns}
\title{Filter Out A-Z Campaigns (protoype).}
\usage{
campaigns(ip, cran_log, case.sensitive = FALSE, min.obs = 5,
  output = "rownames")
}
\arguments{
\item{ip}{Numeric. Nominal IP address (ip_id).}

\item{cran_log}{Object. Package log entries.}

\item{case.sensitive}{Logical.}

\item{min.obs}{Numeric. Threshold number of downloads.}

\item{output}{Character. "df" or "rownames".}
}
\description{
Uses run length encoding rle().
}
\note{
For use with ipFilter3().
}
