% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{campaigns2}
\alias{campaigns2}
\title{Filter Out A-Z Campaigns (protoype).}
\usage{
campaigns2(cran_log, no.of.ips = 100, rle.depth = 100,
  case.sensitive = FALSE, multi.core = TRUE)
}
\arguments{
\item{cran_log}{Object. Package log entries.}

\item{no.of.ips}{Numeric. Number of IP addresses to check.}

\item{rle.depth}{s Numeric. Ceiling for number of rows of run length encoding.}

\item{case.sensitive}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Uses run length encoding rle().
}
