% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check10CharDate.R
\name{check10CharDate}
\alias{check10CharDate}
\title{Check and validate "yyyy-mm-dd" date.}
\usage{
check10CharDate(date, repository = "CRAN")
}
\arguments{
\item{date}{Character. \code{"yyyy-mm-dd"}.}

\item{repository}{Character. "CRAN" or "MRAN".}
}
\description{
Check and validate "yyyy-mm-dd" date.
}
