% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackage.R
\name{checkPackage}
\alias{checkPackage}
\title{Check for valid package names.}
\usage{
checkPackage(packages, dev.mode = FALSE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}
}
\description{
"spell check" package names.
}
