% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockTime.R
\name{clockTime}
\alias{clockTime}
\title{Nominal time to clock time}
\usage{
clockTime(time = "09:00:00", date = "2020-08-14", tz = "Europe/Vienna")
}
\arguments{
\item{time}{Character. "hh:mmm:ss".}

\item{date}{Character. "yyyy-mm-dd".}

\item{tz}{Character. Time zone.}
}
\description{
Time via POSIXlt date.
}
