% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryDistribution.R
\name{countryDistribution}
\alias{countryDistribution}
\title{Tabulate package downloads by country.}
\usage{
countryDistribution(date = Sys.Date() - 1, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
