% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran.dependencies}
\alias{cran.dependencies}
\title{CRAN package dependencies.}
\format{An object of class \code{list} of length 10.}
\usage{
cran.dependencies
}
\description{
CRAN package dependencies.
}
\keyword{datasets}
