% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranData.R
\name{cranData}
\alias{cranData}
\title{Scrape CRAN package information.}
\usage{
cranData(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
Current version, date and size (source).
}
