% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDependencies.R
\name{cranDependencies}
\alias{cranDependencies}
\title{Extract package dependencies from CRAN.}
\usage{
cranDependencies(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Extract package dependencies from CRAN.
}
\note{
Documents cran.dependencies. Uses 'miniCRAN' package.
}
