% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDependenciesNetwork.R
\name{cranDependenciesNetwork}
\alias{cranDependenciesNetwork}
\title{Package dependencies graph.}
\usage{
cranDependenciesNetwork(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}
}
\description{
Package dependencies graph.
}
\note{
Uses 'igraph' package.
}
