% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDistribution.R
\name{cranDistribution}
\alias{cranDistribution}
\title{CRAN distribution (prototype).}
\usage{
cranDistribution(date = NULL, all.filters = FALSE, ip.filter = FALSE,
  small.filter = FALSE, memoization = TRUE, multi.core = FALSE)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{small.filter}{Logical. TRUE filters out downloads less than 1000 bytes.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From Posit's CRAN Mirror http://cran-logs.rstudio.com/
}
