% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloads.R
\name{cranDownloads}
\alias{cranDownloads}
\title{Daily package downloads from the RStudio CRAN mirror.}
\usage{
cranDownloads(packages = NULL, when = c("last-day", "last-week",
  "last-month"), start.year = NULL, start.month = NULL,
  start.day = NULL, end.year = NULL, end.month = NULL,
  end.day = NULL)
}
\arguments{
\item{packages}{A character vector, the packages to query,
or \code{NULL} for a sum of downloads for all packages.
Alternatively, it can also be \code{"R"}, to query downloads
of R itself. \code{"R"} cannot be mixed with packages.}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month}.
If this is given, then \code{from} and \code{to} are ignored.}

\item{start.year}{Numeric.}

\item{start.month}{Numeric.}

\item{start.day}{Numeric.}

\item{end.year}{Numeric.}

\item{end.month}{Numeric.}

\item{end.day}{Numeric.}
}
\description{
S3 implementation of cranlogs::cran_downloads().
}
