% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloadsB.R
\name{cranDownloadsB}
\alias{cranDownloadsB}
\title{Daily package downloads from the RStudio CRAN mirror.}
\usage{
cranDownloadsB(packages = NULL, when = NULL, from = NULL, to = NULL,
  fix.cranlogs = TRUE)
}
\arguments{
\item{packages}{A character vector, the packages to query,
or \code{NULL} for a sum of downloads for all packages.
Alternatively, it can also be \code{"R"}, to query downloads
of R itself. \code{"R"} cannot be mixed with packages.}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month}.
If this is given, then \code{from} and \code{to} are ignored.}

\item{from}{Start date as \code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}.}

\item{to}{End date as \code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}.}

\item{fix.cranlogs}{Logical. Use RStudio logs to fix 8 dates with duplicated data in 'cranlogs' results.}
}
\description{
Enhanced implementation of cranlogs::cran_downloads().
}
