% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflationPlot.R
\name{cranDownloadsPlot}
\alias{cranDownloadsPlot}
\title{CRAN Downloads Plot}
\usage{
cranDownloadsPlot(dataset = "october")
}
\arguments{
\item{dataset}{Character. "october" or "july" for October 2019 or July 2020.}
}
\description{
Document code.
}
