% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cranFilterCounts}
\alias{cranFilterCounts}
\title{CRAN Filter Counts.}
\usage{
cranFilterCounts(lst, multi.core = TRUE)
}
\arguments{
\item{lst}{Object. cran_log list of data frames.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
CRAN Filter Counts.
}
