% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranMirrors.R
\name{cranMirrors}
\alias{cranMirrors}
\title{Scrape CRAN Mirrors data.}
\usage{
cranMirrors(mirror.description = FALSE)
}
\arguments{
\item{mirror.description}{Logical. Mirror details.}
}
\description{
https://cran.r-project.org/mirrors.html
}
