% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranStatistics.R
\name{cranStatistics}
\alias{cranStatistics}
\title{CRAN statistics}
\usage{
cranStatistics(date = Sys.Date() - 1, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date YYYY-MM-DD.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
CRAN statistics
}
