% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{currentTime}
\alias{currentTime}
\title{Compute Current Local Time in Selected Time Zone.}
\usage{
currentTime(tz = "Australia/Sydney")
}
\arguments{
\item{tz}{Character. Local time zone. See OlsonNames().}
}
\description{
Compute Current Local Time in Selected Time Zone.
}
