% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTime.R
\name{dateTime}
\alias{dateTime}
\title{Convert Nominal Date and Time to Date-Time Object.}
\usage{
dateTime(date = "2021-01-01", time = "09:00", tz = "GMT")
}
\arguments{
\item{date}{Character. "yyyy-mm-dd".}

\item{time}{Character. "hh:mm" or "hh:mm:ss".}

\item{tz}{Character. Time zone.}
}
\description{
Character(s) to POSIXlt date.
}
