% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCountry.R
\name{downloadsCountry}
\alias{downloadsCountry}
\title{Compute Downloads by Country Code.}
\usage{
downloadsCountry(month_cran_log, multi.core = FALSE)
}
\arguments{
\item{month_cran_log}{Object.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores to use. Note that due to performance considerations, the number of cores defaults to one on Windows.}
}
\description{
Compute Downloads by Country Code.
}
