% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyFolderAudit.R
\name{emptyFolderAudit}
\alias{emptyFolderAudit}
\title{Check for empty folders CRAN's archive.}
\usage{
emptyFolderAudit(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores}
}
\description{
Empty folders are errors that do not represent real packages.
}
