% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesFunctions.R
\name{extractArchiveDate}
\alias{extractArchiveDate}
\title{Extract a package's archive date.}
\usage{
extractArchiveDate(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
An R data frame.
}
\description{
Date a package is moved to Archive (if available).
}
