% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLog.R
\name{fetchLog2}
\alias{fetchLog2}
\title{Get gzipped data at URL.}
\usage{
fetchLog2(x, memoization = FALSE)
}
\arguments{
\item{x}{Character. URL.}

\item{memoization}{Logical. Use memoization}
}
\description{
Base R version.
}
