% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLog.R
\name{fetchLogBase}
\alias{fetchLogBase}
\title{Get gzipped data at URL.}
\usage{
fetchLogBase(x)
}
\arguments{
\item{x}{Character. URL.}
}
\description{
Get gzipped data at URL.
}
\note{
mfetchLogBase() is memoized version.
}
