% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredDownloads.R
\name{filteredDownloads}
\alias{filteredDownloads}
\title{Filtered package downloads from the RStudio CRAN mirror (prototype).}
\usage{
filteredDownloads(packages = "HistData", date = NULL, all.filters = TRUE,
  ip.filter = FALSE, triplet.filter = FALSE, small.filter = FALSE,
  sequence.filter = FALSE, size.filter = FALSE, check.package = TRUE,
  memoization = TRUE, multi.core = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{small.filter}{Logical. TRUE filters out downloads less than 1000 bytes.}

\item{sequence.filter}{Logical.}

\item{size.filter}{Logical.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
ip, triplet, small, sequence and size filters.
}
