% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDate_2012.R
\name{fixDate_2012}
\alias{fixDate_2012}
\title{Re-map filenames (dates) for 2012 download logs.}
\usage{
fixDate_2012(date = "2012-12-31")
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd".}
}
\value{
A one unit R date or character vector.
}
\description{
Correct for mis-labeled filenames for 2012 logs at RStudio's CRAN Mirror http://cran-logs.rstudio.com/.
}
\note{
This date problem does not affect \code{cranDownlaods()}.
}
