% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySequences.R
\name{identifySequences}
\alias{identifySequences}
\title{Extract Archive sequences from logs (prototype).}
\usage{
identifySequences(pkg.data, arch.pkg.history, download.time = 30)
}
\arguments{
\item{pkg.data}{Object.}

\item{arch.pkg.history}{Object.}

\item{download.time}{Numeric. Package download time allowance (seconds).}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
