% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySequences0.R
\name{identifySequences0}
\alias{identifySequences0}
\title{Extract Archive sequences from logs (prototype).}
\usage{
identifySequences0(package = "cholera", date = Sys.Date() - 1,
  download.time = 30, triplet.filter = TRUE, small.filter = TRUE,
  memoization = TRUE, dev.mode = FALSE)
}
\arguments{
\item{package}{Character. Package name.}

\item{date}{Character. Date.}

\item{download.time}{Numeric. Package download time allowance (seconds).}

\item{triplet.filter}{Logical.}

\item{small.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
