% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTriplets.R
\name{identifyTriplets}
\alias{identifyTriplets}
\title{Extract triplets from logs (prototype).}
\usage{
identifyTriplets(dat, output = "data.frame", time.window = 2,
  time.sort = TRUE, multi.core = TRUE)
}
\arguments{
\item{dat}{Object. CRAN log data frame}

\item{output}{Character. "list" or "data.frame".}

\item{time.window}{Numeric. Seconds.}

\item{time.sort}{Logical. Sort output by time.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
