% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflationPlot.R
\name{inflationPlot2}
\alias{inflationPlot2}
\title{Inflation plots of effects of "small" downloads on aggregate CRAN downloads for October 2019 and July 2020.}
\usage{
inflationPlot2(dataset = "october", filter = "small", wed = FALSE,
  subtitle = TRUE, legend.loc = "topleft")
}
\arguments{
\item{dataset}{Character. "october" or "july" for October 2019 or July 2020.}

\item{filter}{Character. "small", "ip", or "ip.small".}

\item{wed}{Logical.}

\item{subtitle}{Logical.}

\item{legend.loc}{Character. Location of legend.}
}
\description{
Document code.
}
