% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ip_filter}
\alias{ip_filter}
\title{Identify IP's that are mirroring CRAN (k-means helper prototype).}
\usage{
ip_filter(cran_log, centers = 2L, nstart = 25L)
}
\arguments{
\item{cran_log}{Object. cran log.}

\item{centers}{Numeric. Number of k's for k-means clustering.}

\item{nstart}{Numeric. Number of random sets.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
