% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter}
\alias{ipFilter}
\title{Identify IP's that are mirroring CRAN (prototype).}
\usage{
ipFilter(date = Sys.Date() - 1, cutpoint = 15000L, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date. yyyy-mm-dd.}

\item{cutpoint}{Numeric. Threshold number of unique packages downloaded.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
