% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter3}
\alias{ipFilter3}
\title{Identify IP's that are mirroring CRAN (k-means helper prototype).}
\usage{
ipFilter3(cran_log, output = "ip", centers = 2L, nstart = 25L,
  memoization = TRUE)
}
\arguments{
\item{cran_log}{Object. cran log.}

\item{output}{Character. "ip" vector of ip address; "df" data frame.}

\item{centers}{Numeric. Number of k's for k-means clustering.}

\item{nstart}{Numeric. Number of random sets.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
