% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipPackage.R
\name{ipPackage}
\alias{ipPackage}
\title{Tabulate an IP's package downloads (prototype).}
\usage{
ipPackage(ip = 10, date = Sys.Date() - 1, memoization = TRUE,
  sort = TRUE, small.filter = FALSE, triplet.filter = FALSE,
  multi.core = TRUE)
}
\arguments{
\item{ip}{Numeric. ip_id.}

\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical. Sort by download count.}

\item{small.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
ip = 10 is a tw top-level domain on 2020-07-09.
}
