% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansCutpoint.R
\name{kmeansCutpoint}
\alias{kmeansCutpoint}
\title{k-means cutpoint.}
\usage{
kmeansCutpoint(lst, i = 1, centers = 2L, nstart = 25L)
}
\arguments{
\item{lst}{Object. List of download log dataframes.}

\item{i}{Numeric. Day/ID.}

\item{centers}{Numeric. Number of k's for k-means clustering.}

\item{nstart}{Numeric. Number of random sets.}
}
\description{
k-means cutpoint.
}
