% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDate.R
\name{logDate}
\alias{logDate}
\title{Compute Effective CRAN Log Date Based on Local and UTC Time (prototype).}
\usage{
logDate(upload.time = "17:00:00")
}
\arguments{
\item{upload.time}{Character. UTC upload time "hh:mm" or "hh:mm:dd".}
}
\value{
An R date object.
}
\description{
RStudio CRAN Mirror Logs for previous day are posted at 17:00:00 UTC.
}
