% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDate.R
\name{logDate0}
\alias{logDate0}
\title{Compute Effective CRAN Log Date Based on Local and UTC Time (manual prototype).}
\usage{
logDate0(date = Sys.Date() + 1, tz.time = "16:05:00",
  tz = "Australia/Sydney", upload.time = "17:00:00", stop.msg = TRUE)
}
\arguments{
\item{date}{Character. Local date "yyyy-mm-dd".}

\item{tz.time}{Character. Local time ime "hh:mm" or "hh:mm:dd".}

\item{tz}{Character. Local time zone.}

\item{upload.time}{Character. UTC upload time "hh:mm" or "hh:mm:dd".}

\item{stop.msg}{Logical. TRUE uses stop(); FALSE uses warning() and returns previous available date.}
}
\value{
An R date object.
}
\description{
RStudio CRAN Mirror Logs for previous day are posted at 17:00:00 UTC.
}
