% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDate2.R
\name{logDate2}
\alias{logDate2}
\title{Compute Effective CRAN Log Date Based on Local and UTC Time (prototype).}
\usage{
logDate2(upload.time = "17:00", warning.msg = TRUE)
}
\arguments{
\item{upload.time}{Character. UTC upload time for logs "hh:mm" or "hh:mm:ss".}

\item{warning.msg}{Logical. TRUE uses warning() if the fuction returns the previous available date.}
}
\value{
An R date object.
}
\description{
RStudio CRAN Mirror Logs for previous day are posted at 17:00:00 UTC.
}
