% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{logInfo}
\alias{logInfo}
\title{Compute Date and Time of Next and Currently Availabe Log.}
\usage{
logInfo(tz = Sys.timezone())
}
\arguments{
\item{tz}{Character. Local time zone. See OlsonNames() or use Sys.timezone().}
}
\description{
GMT and Local Posting Times.
}
