% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{logPostInfo}
\alias{logPostInfo}
\title{Compute Date of log, GMT and Local Posing Time.}
\usage{
logPostInfo(tz = Sys.timezone())
}
\arguments{
\item{tz}{Character. Local time zone. Use OlsonNames() or Sys.timezone().}
}
\description{
Compute Date of log, GMT and Local Posing Time.
}
