% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing.dates}
\alias{missing.dates}
\title{Missing/NA Posit/RStudio logs.}
\format{
An object of class \code{Date} of length 7.
}
\usage{
missing.dates
}
\description{
August 25-26, 29-31 2025; September 1-2, 2025. 7 days.
}
\keyword{datasets}
