% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersionPercent.R
\name{monthlyLog}
\alias{monthlyLog}
\title{Get CRAN logs for selected month.}
\usage{
monthlyLog(yr.mo = "2020-07")
}
\arguments{
\item{yr.mo}{Character. "yyyy-mm".}
}
\description{
Compute list of log files, 'lst', for packageVersionPercent().
}
\note{
This is computationally intensive; you're downloading 30 odd files that are each around 50 MB in size (and creating a ~1.5 GB file)! Parallelization not practical; multiple attempts to connect to website causes problems. Truncates in-progress/future dates to yesterday's date. Automatically takes care of leap days (e.g., monthlyLog("2020-02").
}
