% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersionPercent.R
\name{monthlyLogs}
\alias{monthlyLogs}
\title{Get CRAN logs for October 2019 or July 2020.}
\usage{
monthlyLogs(month = "October")
}
\arguments{
\item{month}{Character. "October" or "July".}
}
\description{
Compute 'lst' for packageVersions().
}
\note{
This is an intensive task since you're downloading 30 odd log files that are each around 50 MB in size.
}
