% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranPackages.R
\name{mranPackages}
\alias{mranPackages}
\title{Extract package data from MRAN (protoype).}
\usage{
mranPackages(date = Sys.Date() - 1, type = "source")
}
\arguments{
\item{date}{Character.}

\item{type}{Character. "source", "mac", or "win".}
}
\description{
Binary or source size.
}
