% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCountry.R
\name{packageCountry}
\alias{packageCountry}
\title{Package download counts by country (prototype).}
\usage{
packageCountry(packages = "HistData", date = Sys.Date() - 1,
  memoization = TRUE, sort = FALSE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
