% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCRAN2.R
\name{packageCRAN2}
\alias{packageCRAN2}
\title{Scrape package data from CRAN.}
\usage{
packageCRAN2(package = "cholera", check.package = TRUE, size = TRUE,
  r.ver = "4.0", multi.core = TRUE)
}
\arguments{
\item{package}{Character. Package name.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{size}{Logical. Include size of source file.}

\item{r.ver}{Character. Current R version; used in directory path.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame or NULL.
}
\description{
Version, date and size (source file) of most recent publication.
}
