% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDepends.R
\name{packageDepends}
\alias{packageDepends}
\title{Extract package Depends from CRAN.}
\usage{
packageDepends(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Extract package Depends from CRAN.
}
