% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDistribution.R
\name{packageDistribution}
\alias{packageDistribution}
\title{Package Download Distribution.}
\usage{
packageDistribution(package = "HistData", date = NULL,
  size.filter = FALSE, memoization = TRUE, check.package = TRUE)
}
\arguments{
\item{package}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd".  NULL uses latest available log.}

\item{size.filter}{Logical. If Logical, TRUE filters out downloads less than 1000 bytes.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}
}
\description{
Package Download Distribution.
}
