% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{packageFilterCounts}
\alias{packageFilterCounts}
\title{Package Filter Counts.}
\usage{
packageFilterCounts(lst, pkg = "cholera", ip.filter = "campaign",
  multi.core = TRUE)
}
\arguments{
\item{lst}{Object. cran_log list of data frames.}

\item{pkg}{Character.}

\item{ip.filter}{Character. "campaign" or "ip".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Package Filter Counts.
}
