% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory2.R
\name{packageHistory2}
\alias{packageHistory2}
\title{Extract package version history CRAN and Archive (prototype).}
\usage{
packageHistory2(package = "cholera", short.date = TRUE)
}
\arguments{
\item{package}{Character. Package name.}

\item{short.date}{Logical}
}
\description{
Date and version of all publications.
}
