% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistoryB.R
\name{packageHistoryB}
\alias{packageHistoryB}
\title{Extract package or R version history.}
\usage{
packageHistoryB(package = "cholera", check.package = TRUE)
}
\arguments{
\item{package}{Character. Vector of package names (including "R").}

\item{check.package}{Logical. Validate and "spell check" package.}
}
\description{
Date and version of all publications.
}
