% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageImports.R
\name{packageImports}
\alias{packageImports}
\title{Extract package Imports from CRAN.}
\usage{
packageImports(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Extract package Imports from CRAN.
}
