% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageImports.R
\name{packageImports2}
\alias{packageImports2}
\title{Extract package Imports from CRAN.}
\usage{
packageImports2()
}
\description{
Extract package Imports from CRAN.
}
\note{
Documents cran.dependencies.
}
