% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{packageLog}
\alias{packageLog}
\title{Get Package Download Logs.}
\usage{
packageLog(packages = NULL, date = Sys.Date() - 1, triplet.filter = TRUE,
  ip.filter = TRUE, small.filter = TRUE, sequence.filter = TRUE,
  memoization = TRUE, check.package = TRUE, dev.mode = FALSE,
  clean.cran_log = TRUE, clean.output = FALSE, multi.core = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{triplet.filter}{Logical.}

\item{ip.filter}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}

\item{clean.cran_log}{Logical.}

\item{clean.output}{Logical. NULL row names.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
