% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{packageLog0}
\alias{packageLog0}
\title{Get Package Download Logs.}
\usage{
packageLog0(packages = "cholera", date = NULL, check.package = TRUE,
  memoization = TRUE, dev.mode = FALSE, clean.output = FALSE,
  multi.core = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical.}

\item{clean.output}{Logical. NULL row names.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
