% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageMRAN.R
\name{packageMRAN}
\alias{packageMRAN}
\title{Extract package data from MRAN (protoype).}
\usage{
packageMRAN(package = "cholera", date = Sys.Date() - 1, type = "source",
  check.package = TRUE)
}
\arguments{
\item{package}{Character. Package name.}

\item{date}{Character.}

\item{type}{Character. "source", "mac", or "win".}

\item{check.package}{Logical. Validate and "spell check" package.}
}
\description{
Binary or source size.
}
