% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagePublished.R
\name{packagePublished}
\alias{packagePublished}
\title{Extract package Published from CRAN.}
\usage{
packagePublished(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Date of most recent version.
}
