% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{packageRank0}
\alias{packageRank0}
\title{Package download counts and rank percentiles.}
\usage{
packageRank0(packages = "HistData", date = NULL, memoization = TRUE,
  check.package = FALSE, dev.mode = FALSE, multi.core = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
