% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesFunctions.R
\name{packages_in_Archive}
\alias{packages_in_Archive}
\title{Packages currently in Archive}
\usage{
packages_in_Archive(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R vector.
}
\description{
Include inactive (retired) packages and previous versions of active packages.
}
