% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesFunctions.R
\name{packages_observed_in_logs}
\alias{packages_observed_in_logs}
\title{Packages observed in download logs.}
\usage{
packages_observed_in_logs(date = NULL)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}
}
\value{
An R vector.
}
\description{
Packages observed in download logs.
}
