% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesFunctions.R
\name{packages_on_CRAN}
\alias{packages_on_CRAN}
\title{Packages currently on CRAN.}
\usage{
packages_on_CRAN()
}
\value{
An R data frame.
}
\description{
Package name, version and date of publication.
}
