% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSample.R
\name{packageSample}
\alias{packageSample}
\title{Stratified random sample of packages.}
\usage{
packageSample(cran_log, sample.pct = 1, multi.core = TRUE)
}
\arguments{
\item{cran_log}{Object. CRAN log.}

\item{sample.pct}{Numeric.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
