% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSample2.R
\name{packageSample2}
\alias{packageSample2}
\title{Stratified random sample of packages for versionPlot().}
\usage{
packageSample2(lst, repository = "cran", strata.samples = 20,
  package.samples = 100, use.seed = TRUE, multi.core = TRUE)
}
\arguments{
\item{lst}{Object. List of CRAN download logs data frames.}

\item{repository}{Character. "cran" or "archive".}

\item{strata.samples}{Numeric. Number of samples from each stratum.}

\item{package.samples}{Numeric. Number of packages to sample from across strata for use in versionPlot().}

\item{use.seed}{Logical. Use todays's date as seed.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
July benchmarks: cran = 61.684; archive = 35.597.
}
