% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesImport.R
\name{packagesImport}
\alias{packagesImport}
\title{Extract package Imports from CRAN.}
\usage{
packagesImport(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Extract package Imports from CRAN.
}
