% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesPublishedOnDate.R
\name{packagesPublishedOnDate}
\alias{packagesPublishedOnDate}
\title{Packages published on given date (prototype).}
\usage{
packagesPublishedOnDate(date = "2019-10-01")
}
\arguments{
\item{date}{\code{yyyy-mm-dd}.}
}
\description{
Uses tools::CRAN_package_db().
}
\note{
check "2019-01-01"
}
