% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersionPercent.R
\name{packageVersionPercent}
\alias{packageVersionPercent}
\title{Compute data for versionPlot().}
\usage{
packageVersionPercent(lst, resample = FALSE, multi.core = TRUE)
}
\arguments{
\item{lst}{Object. List of CRAN download logs data frames. Use monthlyLogs().}

\item{resample}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
packageRank::blog.data or recompute random sample of packages.
}
