% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersions.R
\name{packageVersions}
\alias{packageVersions}
\title{Extract package version history CRAN and Archive (prototype).}
\usage{
packageVersions(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Date and version of most recent publication.
}
