% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairEntry}
\alias{pairEntry}
\title{Extract small downloads pairs (prototype).}
\usage{
pairEntry(id, x)
}
\arguments{
\item{id}{Character. pairLogID() output.}

\item{x}{Object. Package log.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
