% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pkgLog}
\alias{pkgLog}
\title{Extract Package Logs.}
\usage{
pkgLog(i = 1, pkg = "cholera", dat = july_log)
}
\arguments{
\item{i}{Numeric. Day/ID.}

\item{pkg}{Character.}

\item{dat}{Object. List of logs.}
}
\description{
Extract Package Logs.
}
