% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{plot.bioconductor}
\alias{plot.bioconductor}
\title{Plot method for bioconductor_downloads().}
\usage{
\method{plot}{bioconductor}(x, count = "download", add.points = TRUE,
  smooth = FALSE, smooth.f = 2/3, ...)
}
\arguments{
\item{x}{object.}

\item{count}{Character. "download" or "ip".}

\item{add.points}{Logical. Add points.}

\item{smooth}{Logical. Add stats::lowess smoother.}

\item{smooth.f}{Numeric. smoother span.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for bioconductor_downloads().
}
\examples{
plot(bioconductorDownloads(pkg = "graph"))
plot(bioconductorDownloads(pkg = "graph", year = 2019))
plot(bioconductorDownloads(pkg = "graph", year = 2014, end.year = 2018, month = 6, end.month = 3))
}
