% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorRank.R
\name{plot.bioconductorRank}
\alias{plot.bioconductorRank}
\title{Plot method for bioconductorRank().}
\usage{
\method{plot}{bioconductorRank}(x, graphics = NULL, log.y = TRUE, ...)
}
\arguments{
\item{x}{An object of class "bioconductor_rank" created by \code{bioconductorRank()}.}

\item{graphics}{Character. "base" or "ggplot2".}

\item{log.y}{Logical. Logarithm of package downloads.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R or ggplot2 plot.
}
\description{
Plot method for bioconductorRank().
}
