% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRankPercentile.R
\name{plot.countRankPercentile}
\alias{plot.countRankPercentile}
\title{Plot method for countRankPercentile()..}
\usage{
\method{plot}{countRankPercentile}(x, type = "histogram", ...)
}
\arguments{
\item{x}{An object of class "countRankPercentile" created by \code{countRankPercentile()}.}

\item{type}{Character. "histogram" or "density".}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for countRankPercentile()..
}
