% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryDistribution.R
\name{plot.countryDistribution}
\alias{plot.countryDistribution}
\title{Plot top 10 package downloads by country domain.}
\usage{
\method{plot}{countryDistribution}(x, ...)
}
\arguments{
\item{x}{An object of class "countryDistribution" created by \code{countryDistribution()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for packageDistribution().
}
