% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsRanks.R
\name{plot.countsRanks}
\alias{plot.countsRanks}
\title{Plot method for countsRanks().}
\usage{
\method{plot}{countsRanks}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for countsRanks().
}
