% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.cranFilterCounts}
\alias{plot.cranFilterCounts}
\title{Plot method for cranFilterCounts().}
\usage{
\method{plot}{cranFilterCounts}(x, filter = "all", smooth = FALSE,
  median = FALSE, legend.loc = "topleft", add.legend = TRUE, ...)
}
\arguments{
\item{x}{object.}

\item{filter}{Character.  "ip", "small", "all".}

\item{smooth}{Logical.}

\item{median}{Logical.}

\item{legend.loc}{Character. Location of legend.}

\item{add.legend}{Logical.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for cranFilterCounts().
}
