% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_downloads2.R
\name{plot.cranlogs}
\alias{plot.cranlogs}
\title{Plot method for cran_downloads2().}
\usage{
\method{plot}{cranlogs}(x, graphics_pkg = "ggplot2", points = TRUE,
  log_y = FALSE, smooth = FALSE, se = FALSE, ...)
}
\arguments{
\item{x}{object.}

\item{graphics_pkg}{Character. "base" or "ggplot2".}

\item{points}{Logical}

\item{log_y}{Logical}

\item{smooth}{Logical.}

\item{se}{Logical.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R or ggplot2 plot.
}
\description{
Plot method for cran_downloads2().
}
