% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.filterCounts}
\alias{plot.filterCounts}
\title{Plot method for filterCounts().}
\usage{
\method{plot}{filterCounts}(x, filter = "all", smooth = FALSE, legend.loc = "topleft", ...)
}
\arguments{
\item{x}{object.}

\item{filter}{Character. "triplet", "ip", "small", "sequence", "all".}

\item{smooth}{Logical.}

\item{legend.loc}{Character. Location of legend.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for filterCounts().
}
