% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansCutpoint.R
\name{plot.kmeansCutpoint}
\alias{plot.kmeansCutpoint}
\title{Plot method for kmeansCutpoint().}
\usage{
\method{plot}{kmeansCutpoint}(x, ...)
}
\arguments{
\item{x}{An object of class "kmeansCutpoint" created by \code{kmeansCutpoint()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for kmeansCutpoint().
}
