% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{plot.package_rank}
\alias{plot.package_rank}
\title{Plot method for packageRank().}
\usage{
\method{plot}{package_rank}(x, graphics_pkg = "ggplot2",
  log_count = TRUE, ...)
}
\arguments{
\item{x}{An object of class "package_rank" created by \code{packageRank()}.}

\item{graphics_pkg}{Character. "base" or "ggplot2".}

\item{log_count}{Logical. Logarithm of package downloads.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R or ggplot2 plot.
}
\description{
Plot method for packageRank().
}
\examples{
\donttest{

plot(packageRank(packages = "HistData", date = "2019-01-01"))
plot(packageRank(packages = c("h2o", "Rcpp", "rstan"), date = "2019-01-01"))
}
}
