% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDistribution.R
\name{plot.packageDistribution}
\alias{plot.packageDistribution}
\title{Plot method for packageDistribution().}
\usage{
\method{plot}{packageDistribution}(x, ...)
}
\arguments{
\item{x}{An object of class "packageDistribution" created by \code{packageDistribution()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for packageDistribution().
}
