% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.packageFilterCounts}
\alias{plot.packageFilterCounts}
\title{Plot method for packageFilterCounts().}
\usage{
\method{plot}{packageFilterCounts}(x, filter = "all", smooth = FALSE,
  median = FALSE, legend.loc = "topleft", ...)
}
\arguments{
\item{x}{object.}

\item{filter}{Character. "triplet", "ip", "small", "sequence", "all".}

\item{smooth}{Logical.}

\item{median}{Logical.}

\item{legend.loc}{Character. Location of legend.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for packageFilterCounts().
}
