% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{plot.packageLog}
\alias{plot.packageLog}
\title{Plot method for packageLog().}
\usage{
\method{plot}{packageLog}(x, type = "1D", time.unit = "second",
  points = TRUE, same.xy = TRUE, ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "1D" or "2D".}

\item{time.unit}{Character. "second", "minute", or "hour".}

\item{points}{Logical. For "hour" and "minute" in 2D plots.}

\item{same.xy}{Logical. Use same scale for multiple packages.}

\item{...}{Additional parameters.}
}
\description{
Plot method for packageLog().
}
