% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{plot.packageRankTime}
\alias{plot.packageRankTime}
\title{Plot method for packageRankTime().}
\usage{
\method{plot}{packageRankTime}(x, graphics = NULL, log_count = TRUE,
  smooth = TRUE, sample_smooth = TRUE, f = 1/3, ...)
}
\arguments{
\item{x}{Object. An object of class "packageRankTime" created by \code{packageRankTime()}.}

\item{graphics}{Character. "base" or "ggplot2".}

\item{log_count}{Logical. Logarithm of package downloads.}

\item{smooth}{Logical. Add smoother for selected package.}

\item{sample_smooth}{Logical. lowess background.}

\item{f}{Numeric. stats::lowess() smoother window. For use with graphics = "base" only.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R or ggplot2 plot.
}
\description{
Plot method for packageRankTime().
}
\examples{
\donttest{
plot(packageRankTime(packages = "HistData", when = "last-week"))
plot(packageRankTime(packages = c("Rcpp", "rlang", "data.table"), when = "last-month"))
}
}
