% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersionPercent.R
\name{plot.packageVersionPercent}
\alias{plot.packageVersionPercent}
\title{Plot method for packageVersionPercent().}
\usage{
\method{plot}{packageVersionPercent}(x, ...)
}
\arguments{
\item{x}{An object of class "packageVersions" created by \code{packageVersions()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for packageVersionPercent().
}
