% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentileRank.R
\name{plot.percentileRank}
\alias{plot.percentileRank}
\title{Plot method for percentileRank()..}
\usage{
\method{plot}{percentileRank}(x, type = "histogram", ...)
}
\arguments{
\item{x}{An object of class "percentileRank" created by \code{percentileRank()}.}

\item{type}{Character. "histogram" or "density".}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for percentileRank()..
}
