% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyDownloads.R
\name{plot.weeklyDownloads}
\alias{plot.weeklyDownloads}
\title{Plot method for annualDownloads().}
\usage{
\method{plot}{weeklyDownloads}(x, nrow = 3L, ...)
}
\arguments{
\item{x}{object.}

\item{nrow}{Numeric. Number of rows for ggplot2 facets.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for annualDownloads().
}
