% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCountry.R
\name{plotDownloadsCountry}
\alias{plotDownloadsCountry}
\title{Plot Compute Downloads by Country Code.}
\usage{
plotDownloadsCountry(x = packageRank::blog.data$download.country)
}
\arguments{
\item{x}{Object.}
}
\description{
Plot Compute Downloads by Country Code.
}
