% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationPlot.R
\name{populationPlot}
\alias{populationPlot}
\title{Visualize a Package's Downloads Relative to "All" CRAN packages over Time.}
\usage{
populationPlot(x, graphics = NULL, log.count = TRUE, smooth = TRUE,
  sample.smooth = TRUE, f = 1/3, sample.pct = 5, multi.core = TRUE)
}
\arguments{
\item{x}{object.}

\item{graphics}{Character. NULL, "base" or "ggplot2".}

\item{log.count}{Logical. Logarithm of package downloads.}

\item{smooth}{Logical. Add smoother.}

\item{sample.smooth}{Logical. Add smoother.}

\item{f}{Numeric. stats::lowess() smoother window. For use with graphics = "base" only.}

\item{sample.pct}{Numeric. Percent of packages to sample.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores to use. Note that due to performance considerations, the number of cores defaults to one on Windows.}
}
\description{
Uses a stratified random sample cohort of packages plus top ten packages.
}
