% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryDistribution.R
\name{print.countryDistribution}
\alias{print.countryDistribution}
\title{Print method for countryDistribution().}
\usage{
\method{print}{countryDistribution}(x, N = 10, ...)
}
\arguments{
\item{x}{object.}

\item{N}{Integer. Top N countries.}

\item{...}{Additional parameters.}
}
\description{
Print method for countryDistribution().
}
