% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloads.R
\name{print.cran_downloads}
\alias{print.cran_downloads}
\title{Print method for packageRank().}
\usage{
\method{print}{cran_downloads}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional parameters.}
}
\description{
Print method for packageRank().
}
