% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloads.R
\name{print.cranDownloads}
\alias{print.cranDownloads}
\title{Print method for cranDownloads().}
\usage{
\method{print}{cranDownloads}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional parameters.}
}
\description{
Print method for cranDownloads().
}
