% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_downloads2.R
\name{print.cranlogs}
\alias{print.cranlogs}
\title{Print method for packageRank().}
\usage{
\method{print}{cranlogs}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional parameters.}
}
\description{
Print method for packageRank().
}
