% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{print.package_rank}
\alias{print.package_rank}
\title{Print method for packageRank().}
\usage{
\method{print}{package_rank}(x, ...)
}
\arguments{
\item{x}{An object of class "package_rank" created by \code{packageRank()}}

\item{...}{Additional parameters.}
}
\description{
Print method for packageRank().
}
