% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{print.package_rank_time}
\alias{print.package_rank_time}
\title{Print method for timeSeriesRank().}
\usage{
\method{print}{package_rank_time}(x, ...)
}
\arguments{
\item{x}{An object of class "time_series" created by \code{packageRankTime()}.}

\item{...}{Additional parameters.}
}
\description{
Print method for timeSeriesRank().
}
