% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDistribution.R
\name{print.packageDistribution}
\alias{print.packageDistribution}
\title{Print method for packageDistribution().}
\usage{
\method{print}{packageDistribution}(x, ...)
}
\arguments{
\item{x}{An object of class "packageDistribution" created by \code{packageDistribution()}}

\item{...}{Additional parameters.}
}
\description{
Print method for packageDistribution().
}
