% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{print.packageLog}
\alias{print.packageLog}
\title{Print method for packageLog().}
\usage{
\method{print}{packageLog}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional parameters.}
}
\description{
Print method for packageLog().
}
