% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{print.packageRank}
\alias{print.packageRank}
\title{Print method for packageRank().}
\usage{
\method{print}{packageRank}(x, ...)
}
\arguments{
\item{x}{An object of class "packageRank" created by \code{packageRank()}}

\item{...}{Additional parameters.}
}
\description{
Print method for packageRank().
}
