% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{print.packageRankTime}
\alias{print.packageRankTime}
\title{Print method for packageRankTime().}
\usage{
\method{print}{packageRankTime}(x, ...)
}
\arguments{
\item{x}{An object of class "packageRankTime" created by \code{packageRankTime()}.}

\item{...}{Additional parameters.}
}
\description{
Print method for packageRankTime().
}
