% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveDate.R
\name{resolveDate}
\alias{resolveDate}
\title{Check and resolve date (prototype).}
\usage{
resolveDate(date, type = "from")
}
\arguments{
\item{date}{Date \code{yyyy-mm-dd}.}

\item{type}{Character. type of data "to" or "from".}
}
\description{
Check and resolve date (prototype).
}
