% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveDate.R
\name{resolveFromDate}
\alias{resolveFromDate}
\title{Check and resolve start date (prototype).}
\usage{
resolveFromDate(date)
}
\arguments{
\item{date}{Date \code{yyyy-mm-dd}.}
}
\description{
Check and resolve start date (prototype).
}
