% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveDate.R
\name{resolveToDate}
\alias{resolveToDate}
\title{Check and resolve end date (prototype).}
\usage{
resolveToDate(date)
}
\arguments{
\item{date}{Date \code{yyyy-mm-dd}.}
}
\description{
Check and resolve end date (prototype).
}
