% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLog.R
\name{rLog}
\alias{rLog}
\title{Get R Application Download Logs.}
\usage{
rLog(date = NULL)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses last available log.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
