% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rstudio.logs}
\alias{rstudio.logs}
\title{Eight RStudio Download Logs to Fix Duplicate Logs Errors in 'cranlogs'.}
\format{
\describe{
   \item{\code{date}}{}
   \item{\code{time}}{}
   \item{\code{size}}{}
   \item{\code{r_version}}{}
   \item{\code{r_arch}}{}
   \item{\code{r_os}}{}
   \item{\code{package}}{}
   \item{\code{version}}{}
   \item{\code{country}}{}
   \item{\code{ip_id}}{}
}
}
\usage{
rstudio.logs
}
\description{
October 6-8, 2012; October 11, 2012; December 26-28; and January 1, 20113.
}
\keyword{datasets}
