% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFilter.R
\name{sequenceFilter}
\alias{sequenceFilter}
\title{Filter downloads of full-sized sequential versions (prototype).}
\usage{
sequenceFilter(dat, max.delta.time = 30)
}
\arguments{
\item{dat}{Object.}

\item{max.delta.time}{Numeric. Maximum time interval between downloads (seconds).}
}
\description{
Filter downloads of full-sized sequential versions (prototype).
}
