% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialVersion.R
\name{sequentialVersion}
\alias{sequentialVersion}
\title{Download of sequential versions}
\usage{
sequentialVersion(package = "cholera", date = Sys.Date() - 1,
  small.filter = TRUE, triplet.filter = TRUE, time.frame = 15,
  one.country = FALSE, one.ip = FALSE)
}
\arguments{
\item{package}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd".}

\item{small.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{time.frame}{Numeric. minutes.}

\item{one.country}{Logical.}

\item{one.ip}{Logical.}
}
\description{
Download of sequential versions
}
\examples{
\dontrun{
packageLog(packages = "cholera", date = "2020-06-27")
sequentialVersion(package = "cholera", date = "2020-06-27")
}
}
