% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFilter.R
\name{sizeFilter}
\alias{sizeFilter}
\title{Filter out small downloads (prototype).}
\usage{
sizeFilter(dat)
}
\arguments{
\item{dat}{Object. Package log entries.}
}
\description{
Filter out small downloads (prototype).
}
