% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallFilter.R
\name{smallFilter}
\alias{smallFilter}
\title{Filter out small downloads (hierachical cluster prototype helper).}
\usage{
smallFilter(dat, filter = TRUE)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{filter}{Logical or Numeric. If Logical, TRUE filters out downloads less than 1000 bytes. If Numeric, a positive value sets the minimum download size (in bytes) to consider; a negative value sets the maximum download size to consider.}
}
\description{
Filter out small downloads (hierachical cluster prototype helper).
}
