% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{summary.bioconductor}
\alias{summary.bioconductor}
\title{Summary method for bioconductorDownloads().}
\usage{
\method{summary}{bioconductor}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\description{
Summary method for bioconductorDownloads().
}
