% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorRank.R
\name{summary.bioconductorRank}
\alias{summary.bioconductorRank}
\title{Summary method for bioconductorRank().}
\usage{
\method{summary}{bioconductorRank}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "bioconductor_rank" created by \code{bioconductorRank()}}

\item{...}{Additional parameters.}
}
\description{
Summary method for bioconductorRank().
}
\note{
This is useful for directly accessing the data frame.
}
