% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRankPercentile.R
\name{summary.countRankPercentile}
\alias{summary.countRankPercentile}
\title{Summary method for countRankPercentile().}
\usage{
\method{summary}{countRankPercentile}(object, ...)
}
\arguments{
\item{object}{An object of class "countRankPercentile" created by \code{countRankPercentile()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R vector
}
\description{
Five number summary of download count distribution
}
