% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloads.R
\name{summary.cran_downloads}
\alias{summary.cran_downloads}
\title{Summary method for packageRank().}
\usage{
\method{summary}{cran_downloads}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\description{
Summary method for packageRank().
}
\note{
This is useful for directly accessing the data frame.
}
