% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranStatistics.R
\name{summary.cran_statistics}
\alias{summary.cran_statistics}
\title{Summary method for cranStatistics().}
\usage{
\method{summary}{cran_statistics}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "cran_statistics" created by \code{cranStatistics()}.}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return summary package counts.
}
