% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDistribution.R
\name{summary.cranDistribution}
\alias{summary.cranDistribution}
\title{Summary method for cranDistribution().}
\usage{
\method{summary}{cranDistribution}(object, ...)
}
\arguments{
\item{object}{An object of class "cranDistribution" created by \code{cranDistribution()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R vector
}
\description{
Five number (+ mean) summary of download count distribution
}
