% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloads.R
\name{summary.cranDownloads}
\alias{summary.cranDownloads}
\title{Summary method for cranDownloads().}
\usage{
\method{summary}{cranDownloads}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\description{
Summary method for cranDownloads().
}
\note{
This is useful for directly accessing the data frame.
}
