% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_downloads2.R
\name{summary.cranlogs}
\alias{summary.cranlogs}
\title{Summary method for packageRank().}
\usage{
\method{summary}{cranlogs}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\description{
Summary method for packageRank().
}
\note{
This is useful for directly accessing the data frame.
}
