% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{summary.package_rank}
\alias{summary.package_rank}
\title{Summary method for packageRank().}
\usage{
\method{summary}{package_rank}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "package_rank" created by \code{packageRank()}}

\item{...}{Additional parameters.}
}
\description{
Summary method for packageRank().
}
\note{
This is useful for directly accessing the data frame.
}
