% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{summary.package_rank_time}
\alias{summary.package_rank_time}
\title{Summary method for timeSeriesRank().}
\usage{
\method{summary}{package_rank_time}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "time_series" created by \code{packageRankTime()}.}

\item{...}{Additional parameters.}
}
\description{
Summary method for timeSeriesRank().
}
\note{
This is useful for directly accessing the data frame.
}
