% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDistribution.R
\name{summary.packageDistribution}
\alias{summary.packageDistribution}
\title{Summary method for class "packageDistribution".}
\usage{
\method{summary}{packageDistribution}(object, ...)
}
\arguments{
\item{object}{An object of class "packageDistribution" created by \code{cranDistribution()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R vector
}
\description{
Five number (+ mean) summary of download count distribution
}
