% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{summary.packageRankTime}
\alias{summary.packageRankTime}
\title{Summary method for packageRankTime().}
\usage{
\method{summary}{packageRankTime}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "packageRankTime" created by \code{packageRankTime()}.}

\item{...}{Additional parameters.}
}
\description{
Summary method for packageRankTime().
}
\note{
This is useful for directly accessing the data frame.
}
