% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskview.package.names}
\alias{taskview.package.names}
\title{List of CRAN Task View package names.}
\format{An object of class \code{list} of length 40.}
\usage{
taskview.package.names
}
\description{
List of CRAN Task View package names.
}
\keyword{datasets}
