% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskview.pkgs}
\alias{taskview.pkgs}
\title{List of task view core packages.}
\format{An object of class \code{list} of length 40.}
\usage{
taskview.pkgs
}
\description{
List of task view core packages.
}
\keyword{datasets}
