% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskViewPackages.R
\name{taskViewCountData}
\alias{taskViewCountData}
\title{Get CRAN Task View data.}
\usage{
taskViewCountData(task.view, when = NULL, from = "2019-01",
  to = NULL)
}
\arguments{
\item{task.view}{Character. Short CRAN Task View.}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month}.
If this is given, then \code{from} and \code{to} are ignored.}

\item{from}{Start date as \code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}.}

\item{to}{End date as \code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}.#'}
}
\description{
Get CRAN Task View data.
}
\examples{
\donttest{
taskViewCountData("Bayesian")
}
}
