% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreTaskViewPackages.R
\name{taskviewData}
\alias{taskviewData}
\title{Get taskview data.}
\usage{
taskviewData(taskview)
}
\arguments{
\item{taskview}{Character. Short task view name}
}
\description{
Get taskview data.
}
