% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskViewPackages.R
\name{taskViewPackageNames}
\alias{taskViewPackageNames}
\title{List of CRAN Task View Package Names}
\usage{
taskViewPackageNames(core.packages = FALSE, multi.core = TRUE)
}
\arguments{
\item{core.packages}{Logical. TRUE for core packages; FALSE for all packages.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}
}
\description{
List of CRAN Task View Package Names
}
\note{
Documents \code{taskview.package.names}
}
