% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskViewPackages.R
\name{taskViewPackages}
\alias{taskViewPackages}
\title{Extract CRAN Task View packages}
\usage{
taskViewPackages(url = url, core.packages = TRUE)
}
\arguments{
\item{url}{Character. Task view URL.}

\item{core.packages}{Logical. TRUE for core packages; FALSE for all packages.}
}
\value{
A character vector of package names.
}
\description{
Extract CRAN Task View packages
}
\examples{
\donttest{
# Bayesian Inference
url <- "https://CRAN.R-project.org/view=Bayesian"
taskViewPackages(url)

# Chemometrics and Computational Physics
url <- "https://CRAN.R-project.org/view=ChemPhys"
taskViewPackages(url, core = FALSE)
}
}
