% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreTaskViewPackages.R
\name{taskviewPlot}
\alias{taskviewPlot}
\title{Plot taskview data.}
\usage{
taskviewPlot(taskview)
}
\arguments{
\item{taskview}{Character. Short task view name}
}
\description{
Plot taskview data.
}
