% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeWindow.R
\name{timeWindow}
\alias{timeWindow}
\title{Create a time window for download log files (prototype).}
\usage{
timeWindow(t, window = 1)
}
\arguments{
\item{t}{Character. Time as "hh:mm:ss".}

\item{window}{Numeric. Delta in seconds for time window.}
}
\description{
Create a time window for download log files (prototype).
}
\examples{
timeWindow("00:02:00")
timeWindow("09:12:53", 2)
}
