% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCountry.R
\name{topCountryCodes}
\alias{topCountryCodes}
\title{Compute Top N Downloads by Country Code.}
\usage{
topCountryCodes(month_cran_log, top.n = 5L, multi.core = TRUE)
}
\arguments{
\item{month_cran_log}{Object.}

\item{top.n}{Integer.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores to use. Note that due to performance considerations, the number of cores defaults to one on Windows.}
}
\description{
Compute Top N Downloads by Country Code.
}
