% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripletFilter2.R
\name{tripletFilter2}
\alias{tripletFilter2}
\title{Filter out small downloads triplets (prototype).}
\usage{
tripletFilter2(dat)
}
\arguments{
\item{dat}{Object. Package log entries.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
