% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePackage.R
\name{validatePackage}
\alias{validatePackage}
\title{Check for valid package names.}
\usage{
validatePackage(packages, include.archive = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{include.archive}{Logical. Include archive when validating package. This is computationally expensive becuase it scrapes https://cran.r-project.org/src/contrib/Archive/.}
}
\description{
Check for valid package names.
}
