% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePackage.R
\name{validatePackage0}
\alias{validatePackage0}
\title{Check for valid package names (scrape CRAN).}
\usage{
validatePackage0(packages, check.archive = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{check.archive}{Logical. Include archive when validating package. This is computationally expensive because it scrapes https://cran.r-project.org/src/contrib/Archive/.}
}
\description{
Check for valid package names (scrape CRAN).
}
