% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePackage2.R
\name{validatePackage2}
\alias{validatePackage2}
\title{Check for valid package names.}
\usage{
validatePackage2(packages)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}
}
\description{
Check for valid package names.
}
