% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionsObserved.R
\name{versionsObserved}
\alias{versionsObserved}
\title{Count of package-version downloads.}
\usage{
versionsObserved(package = "cholera", date = Sys.Date() - 1,
  all.history = FALSE, size.filter = FALSE)
}
\arguments{
\item{package}{Character. Package Name}

\item{date}{Character. Date. "yyyy-mm-dd".}

\item{all.history}{Logical. Include entire package history.}

\item{size.filter}{Logical or Numeric. If Logical, TRUE filters out downloads less than 1000 bytes. If Numeric, a postive value sets the minimum download size (in bytes) to consider; a negative value sets the maximum download size to consider.}
}
\description{
Count of package-version downloads.
}
