% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyDownloads.R
\name{weeklyDownloads}
\alias{weeklyDownloads}
\title{Sample Weekly CRAN Downloads Data.}
\usage{
weeklyDownloads(start.yr = 2013, n = 50)
}
\arguments{
\item{start.yr}{Numeric or Integer.}

\item{n}{Numeric or Integer. Number of weeks (samples).}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
