% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualDownloads.R
\name{annualDownloads}
\alias{annualDownloads}
\title{Count Total CRAN Download.}
\usage{
annualDownloads(start.yr = 2013, end.yr = 2023)
}
\arguments{
\item{start.yr}{Numeric or Integer.}

\item{end.yr}{Numeric or Integer.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
A way around Gateway Timeout (HTTP 504). This takes a while since it computes each year separately.s
}
