% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualDownloads.R
\name{annualDownloads}
\alias{annualDownloads}
\title{Count Total CRAN Download.}
\usage{
annualDownloads(start.yr = 2013, end.yr = 2020, multi.core = TRUE)
}
\arguments{
\item{start.yr}{Numeric or Integer.}

\item{end.yr}{Numeric or Integer.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
