% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyTriplet.R
\name{anyTriplet}
\alias{anyTriplet}
\title{Detect small downloads triplets (prototype).}
\usage{
anyTriplet(x, sample.pct = 5, multi.core = TRUE)
}
\arguments{
\item{x}{Object. Package log.}

\item{sample.pct}{Numeric. Percent of packages to sample.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores to use. Note that due to performance considerations, the number of cores defaults to one on Windows.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
