% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivePackages.R
\name{archivePackages}
\alias{archivePackages}
\title{Packages in CRAN archive.}
\usage{
archivePackages(include.date = FALSE)
}
\arguments{
\item{include.date}{Logical. Return data frame with package name and last publication date.}
}
\description{
Packages in CRAN archive.
}
