% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivePackages.R
\name{archivePackages}
\alias{archivePackages}
\title{Packages in CRAN archive.}
\usage{
archivePackages(include.date = FALSE, multi.core = TRUE, dev.mode = FALSE)
}
\arguments{
\item{include.date}{Logical. Return data frame with package name and last publication date.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
Scrape https://cran.r-project.org/src/contrib/Archive/.
}
