% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads2.R
\name{bioconductor_downloads2}
\alias{bioconductor_downloads2}
\title{Annual/monthly package downloads from Bioconductor (prototype).}
\usage{
bioconductor_downloads2(pkg = NULL, year = NULL, month = NULL,
  end.year = NULL, end.month = NULL, observation = "month")
}
\arguments{
\item{pkg}{Character.}

\item{year}{Numeric.}

\item{month}{Numeric.}

\item{end.year}{Numeric.}

\item{end.month}{Numeric.}

\item{observation}{Character. "year" or "month"}
}
\description{
Annual/monthly package downloads from Bioconductor (prototype).
}
\examples{
# entire history
bioconductor_downloads2(pkg = "clusterProfiler")

# year-to-date
bioconductor_downloads2(pkg = "clusterProfiler", year = 2019)

# June 2014 througg March 2018
bioconductor_downloads2(pkg = "clusterProfiler", year = 2014,
  end.year = 2018, month = 6, end.month = 3)

# last 12 months
bioconductor_downloads2(pkg = "clusterProfiler", year = "last-year")
}
