% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads2.R
\name{bioconductor_downloads2}
\alias{bioconductor_downloads2}
\title{Annual/monthly package downloads from Bioconductor (prototype).}
\usage{
bioconductor_downloads2(year = NULL, month = NULL, end.year = NULL,
  end.month = NULL, pkg = NULL, observation = "year")
}
\arguments{
\item{year}{Numeric.}

\item{month}{Numeric.}

\item{end.year}{Numeric.}

\item{end.month}{Numeric.}

\item{pkg}{Character.}

\item{observation}{Character. "year" or "month"}
}
\description{
Annual/monthly package downloads from Bioconductor (prototype).
}
