% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{bioconductorDownloads}
\alias{bioconductorDownloads}
\title{Annual/monthly package downloads from Bioconductor (beta).}
\usage{
bioconductorDownloads(pkg = NULL, year = NULL, month = NULL,
  end.year = NULL, end.month = NULL, observation = "month")
}
\arguments{
\item{pkg}{Character. Vector of package names.}

\item{year}{Numeric.}

\item{month}{Numeric.}

\item{end.year}{Numeric.}

\item{end.month}{Numeric.}

\item{observation}{Character. "year" or "month"}
}
\description{
Annual/monthly package downloads from Bioconductor (beta).
}
\examples{
# all packages
bioconductorDownloads()

# entire history
bioconductorDownloads(pkg = "clusterProfiler")

# year-to-date
bioconductorDownloads(pkg = "clusterProfiler", year = 2019)

# June 2014 througg March 2018
bioconductorDownloads(pkg = "clusterProfiler", year = 2014,
  end.year = 2018, month = 6, end.month = 3)

# last 12 months
bioconductorDownloads(pkg = "clusterProfiler", year = "last-year")
}
