% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{bioconductorDownloads}
\alias{bioconductorDownloads}
\title{Annual/monthly package downloads from Bioconductor (beta).}
\usage{
bioconductorDownloads(packages = NULL, from = NULL, to = NULL,
  when = NULL, observation = "month")
}
\arguments{
\item{packages}{Character. Vector of package names.}

\item{from}{Start date as \code{yyyy-mm} or \code{yyyy}.}

\item{to}{End date as \code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}.}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month}.
If this is given, then \code{from} and \code{to} are ignored.}

\item{observation}{"year" or "month".}
}
\description{
Annual/monthly package downloads from Bioconductor (beta).
}
\examples{
\donttest{
# all packages
bioconductorDownloads()

# entire history
bioconductorDownloads(packages = "clusterProfiler")

# year-to-date
bioconductorDownloads(packages = "clusterProfiler", from = 2014, to = 2015)

# June 2014 througg March 2018
bioconductorDownloads(packages = "clusterProfiler", from = "2014-06", to = "2015-03")

# last 12 months
bioconductorDownloads(packages = "clusterProfiler", when = "last-year")
}
}
