% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorRank.R
\name{bioconductorRank}
\alias{bioconductorRank}
\title{Package download counts and rank percentiles.}
\usage{
bioconductorRank(package = "monocle", date = "2019-01",
  count = "download")
}
\arguments{
\item{package}{Character. Vector of package name(s).}

\item{date}{Character. Date. yyyy-mm}

\item{count}{Character. "ip" or "download".}
}
\value{
An R data frame.
}
\description{
From bioconductor
}
\examples{
\dontrun{
bioconductorRank(package = "cicero", date = "2019-09")
}
}
