% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blog.data}
\alias{blog.data}
\title{Blog post data.}
\format{
A list with 29 elements.
}
\usage{
blog.data
}
\description{
\describe{
   \item{\code{archive.pkg_ver}}{}
   \item{\code{archive.pkg_ver.filtered}}{}
   \item{\code{cran.pkg_ver}}{}
   \item{\code{cran.pkg_ver.filtered}}{}
   \item{\code{dl.ct}}{}
   \item{\code{dl.ct2}}{}
   \item{\code{pkg.ct}}{}
   \item{\code{pkg.ct2}}{}
   \item{\code{oct.data}}{}
   \item{\code{cholera.data}}{}
   \item{\code{ggplot2.data}}{}
   \item{\code{VR.data}}{}
   \item{\code{smpl}}{}
   \item{\code{smpl.histories}}{}
   \item{\code{smpl.archive}}{}
   \item{\code{smpl.archive.histories}}{}
   \item{\code{ccode.ct}}{}
   \item{\code{crosstab_2019_10_01}}{}
   \item{\code{percentiles}}{}
   \item{\code{top.n.oct2019}}{}
   \item{\code{top.n.jul2020}}{}
   \item{\code{download.country}}{}
   \item{\code{october.downloads}}{}
   \item{\code{july.downloads}}{}
   \item{\code{cran.pkgs.oct}}{}
   \item{\code{arch.pkgs.oct}}{}
   \item{\code{cran.pkgs.jul}}{}
   \item{\code{arch.pkgs.jul}}{}
   \item{\code{pkg.history}}{}
}
}
\keyword{datasets}
