% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{campaigns}
\alias{campaigns}
\title{Filter out A-Z campaigns (protoype).}
\usage{
campaigns(ip, cran_log, min.obs = 5, output = "rownames")
}
\arguments{
\item{ip}{Numeric. Nominal IP address.}

\item{cran_log}{Object. Package log entries.}

\item{min.obs}{Numeric. Threshold number of unique packages downloaded.}

\item{output}{Character. "df" or "rownames".}
}
\description{
Uses run length encoding rle().
}
\note{
For use with ipFilter3().
}
