% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{campaigns}
\alias{campaigns}
\title{Filter out A-Z campaigns (protoype).}
\usage{
campaigns(tmp, min.obs = 5)
}
\arguments{
\item{tmp}{Object. Package log entries.}

\item{min.obs}{Numeric. Threshold number of unique packages downloaded.}
}
\description{
Uses run length encoding rle().
}
\note{
For use with ipFilter3().
}
