% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDate.R
\name{checkDate}
\alias{checkDate}
\title{Check and validate "yyyy-mm-dd" date.}
\usage{
checkDate(date, repository = "CRAN")
}
\arguments{
\item{date}{Character. \code{"yyyy-mm-dd"}.}

\item{repository}{Character. "CRAN" or "MRAN".}
}
\description{
Check and validate "yyyy-mm-dd" date.
}
