% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreTaskViewPackages.R
\name{coreTaskViewPackages}
\alias{coreTaskViewPackages}
\title{Extract core CRAN Task View packages}
\usage{
coreTaskViewPackages(url)
}
\value{
A character vector of package names.
}
\description{
url Character. Task view URL.
}
\examples{
# Bayesian Inference
url <- "https://cran.r-project.org/web/views/Bayesian.html"
coreTaskViewPackages(url)

# Chemometrics and Computational Physics
url <- "https://cran.r-project.org/web/views/ChemPhys.html"
coreTaskViewPackages(url)
}
