% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryDistribution.R
\name{countryDistribution}
\alias{countryDistribution}
\title{Tabulate package downloads by country.}
\usage{
countryDistribution(date = NULL, all.filters = FALSE, ip.filter = FALSE,
  sequence.filter = FALSE, size.filter = FALSE, small.filter = FALSE,
  version.filter = FALSE, memoization = TRUE, multi.core = FALSE)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{size.filter}{Logical.}

\item{small.filter}{Logical. TRUE filters out downloads less than 1000 bytes.}

\item{version.filter}{Logical. TRUE selects only most recent version.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
all.filters = TRUE only enables IP and small filters.
}
