% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryPackage.R
\name{countryPackage}
\alias{countryPackage}
\title{Tabulate a country's package downloads.}
\usage{
countryPackage(country = "HK", date = Sys.Date() - 1, memoization = TRUE,
  sort = TRUE, triplet.filter = TRUE, ip.filter = TRUE,
  ip.campaigns = TRUE, small.filter = TRUE, sequence.filter = FALSE,
  multi.core = TRUE)
}
\arguments{
\item{country}{Character. country abbreviation.}

\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical. Sort by download count.}

\item{triplet.filter}{Logical.}

\item{ip.filter}{Logical.}

\item{ip.campaigns}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
"US" outlier 10 min with all filters!
}
