% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryPackage.R
\name{countryPackage}
\alias{countryPackage}
\title{Tabulate a country's package downloads.}
\usage{
countryPackage(country = "HK", date = NULL, all.filters = FALSE,
  ip.filter = FALSE, small.filter = FALSE, sequence.filter = FALSE,
  size.filter = FALSE, sort.count = TRUE, memoization = TRUE)
}
\arguments{
\item{country}{Character. country abbreviation.}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{small.filter}{Logical. TRUE filters out downloads less than 1000 bytes.}

\item{sequence.filter}{Logical. Set to FALSE.}

\item{size.filter}{Logical. Set to FALSE.}

\item{sort.count}{Logical. Sort by download count.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
"US" outlier 10 min with all filters!
}
