% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryPackage.R
\name{countryPackage}
\alias{countryPackage}
\title{Tabulate a country's package downloads.}
\usage{
countryPackage(country = "US", date = Sys.Date() - 1, memoization = TRUE,
  sort = TRUE)
}
\arguments{
\item{country}{Character. country abbreviation.}

\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical. Sort by download count.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
