% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsRanks.R
\name{countsRanks}
\alias{countsRanks}
\title{Counts v. Rank Percentiles.}
\usage{
countsRanks(package = "cholera", size.filter = FALSE, memoization = FALSE)
}
\arguments{
\item{package}{Character.}

\item{size.filter}{Logical.}

\item{memoization}{Logical.}
}
\description{
First week of March 2020.
}
