% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsRanks.R
\name{countsRanks}
\alias{countsRanks}
\title{Counts v. Rank Percentiles for 'cholera' for First Week of March 2020.}
\usage{
countsRanks(package = "cholera", size.filter = FALSE)
}
\arguments{
\item{package}{Character.}

\item{size.filter}{Logical.}
}
\description{
Document code for blog graph.
}
