% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDependencies.R
\name{cranDependencies}
\alias{cranDependencies}
\title{Imports dependencies.}
\usage{
cranDependencies()
}
\description{
Documents cran.dependencies.
}
\note{
2019-09-20 14,960 (14,945); 2019-09-23 14,976 (14,961); 2019-09-26 (14,965)
}
