% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloadsB.R
\name{cranDownloadsB}
\alias{cranDownloadsB}
\title{Filtered package downloads from the RStudio CRAN mirror (protoype).}
\usage{
cranDownloadsB(packages = "HistData", date = Sys.Date() - 1,
  check.package = TRUE, triplet.filter = TRUE, ip.filter = TRUE,
  small.filter = TRUE, sequence.filter = TRUE, memoization = TRUE,
  dev.mode = FALSE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{triplet.filter}{Logical.}

\item{ip.filter}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}
}
\description{
triplet, ip and small filters.
}
