% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDownloadsB.R
\name{cranDownloadsB}
\alias{cranDownloadsB}
\title{Filtered package downloads from the RStudio CRAN mirror. (protoype)}
\usage{
cranDownloadsB(packages = "HistData", date = Sys.Date() - 1,
  check.package = TRUE, small.filter = TRUE, triplet.filter = TRUE,
  memoization = TRUE, dev.mode = FALSE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{small.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}
}
\description{
Filters out small and intermediate downloads.
}
