% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTime.R
\name{dateTime}
\alias{dateTime}
\title{Nominal time to clock time}
\usage{
dateTime(date = "2021-01-01", time = "09:00", tz = "GMT")
}
\arguments{
\item{date}{Character. "yyyy-mm-dd".}

\item{time}{Character. "hh:mmm" or "hh:mmm:ss".}

\item{tz}{Character. Time zone.}
}
\description{
Time via POSIXlt date.
}
