% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLog.R
\name{fetchCranLog}
\alias{fetchCranLog}
\title{Fetch CRAN Logs.}
\usage{
fetchCranLog(date, memoization)
}
\arguments{
\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
Fetch CRAN Logs.
}
