% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLog2.R
\name{fetchLog2}
\alias{fetchLog2}
\title{Fetch Package Logs.}
\usage{
fetchLog2(date, memoization)
}
\arguments{
\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
Fetch Package Logs.
}
