% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySequences.R
\name{identifySequences}
\alias{identifySequences}
\title{Extract Archive sequences from logs (prototype).}
\usage{
identifySequences(package = "cholera", date = Sys.Date() - 1,
  download.time = 30, memoization = TRUE, dev.mode = FALSE)
}
\arguments{
\item{package}{Character. Package name.}

\item{date}{Character. Date.}

\item{download.time}{Numeric. Package download time allowance (seconds).}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
