% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflationPlot.R
\name{inflationPlot}
\alias{inflationPlot}
\title{Inflation plots of effects of "small" downloads and prior versions for October 2019: 'cholera', 'ggplot2', and 'VR'.}
\usage{
inflationPlot(package = "cholera", filter = "size", legend.loc = "topleft")
}
\arguments{
\item{package}{Character.}

\item{filter}{Character. Size, version, or size and version}

\item{legend.loc}{Character. Location of legend.}
}
\description{
Document code for blog graph.
}
