% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipCount.R
\name{ipCount}
\alias{ipCount}
\title{Count number of IP addresses.}
\usage{
ipCount(date = NULL, memoization = TRUE, sort.count = TRUE)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort.count}{Logical. Sort by download count.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
