% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipDownloads.R
\name{ipDownloads}
\alias{ipDownloads}
\title{Package download count  by IP distribution summary.}
\usage{
ipDownloads(date = Sys.Date() - 1, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
