% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipDownloads.R
\name{ipDownloads}
\alias{ipDownloads}
\title{Unique package download counts by IP address.}
\usage{
ipDownloads(date = NULL, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
