% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter}
\alias{ipFilter}
\title{Identify IP's that are mirroring CRAN (prototype).}
\usage{
ipFilter(dat, cutpoint = 15000L)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{cutpoint}{Numeric. Threshold number of unique packages downloaded.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
