% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter}
\alias{ipFilter}
\title{Filter Out A-Z Campaigns from IPs with many unique package downloads.}
\usage{
ipFilter(cran_log, campaigns = TRUE, rle.depth = 100,
  case.sensitive = FALSE, multi.core = TRUE)
}
\arguments{
\item{cran_log}{Object. Package log entries.}

\item{campaigns}{Logical. Filter A-Z campaigns when checking IPs with high unique package download counts.}

\item{rle.depth}{s Numeric. Ceiling for number of rows of run length encoding. Fewer rows means longer runs}

\item{case.sensitive}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Uses run length encoding, rle(), and k-means clustering, stats::kmeans().
}
