% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter0}
\alias{ipFilter0}
\title{Identify IP's that are mirroring CRAN (standalone prototype).}
\usage{
ipFilter0(date = Sys.Date() - 1, cutpoint = 5000L, memoization = TRUE)
}
\arguments{
\item{date}{Character. Date.}

\item{cutpoint}{Numeric. Threshold of unique packages downloaded.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
