% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipFilter.R
\name{ipFilter2}
\alias{ipFilter2}
\title{Identify IP's that are mirroring CRAN (k-means standalone prototype).}
\usage{
ipFilter2(date = Sys.Date() - 1, output = "ip", centers = 2L,
  nstart = 25L, memoization = TRUE, dev.mode = FALSE)
}
\arguments{
\item{date}{Character. Date.}

\item{output}{Character. "ip" vector of ip address; "df" data frame.}

\item{centers}{Numeric. Number of k's for k-means clustering.}

\item{nstart}{Numeric. Number of random sets.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{dev.mode}{Logical. Use fetchLogBase().}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
