% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipPackage.R
\name{ipPackage}
\alias{ipPackage}
\title{Tabulate a IP address's package downloads (prototype).}
\usage{
ipPackage(ip = 1, date = Sys.Date() - 1, memoization = TRUE, sort = TRUE)
}
\arguments{
\item{ip}{Numeric. ip_id (anonymized IP address).}

\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical. Sort by download count.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
