% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipPackage.R
\name{ipPackage}
\alias{ipPackage}
\title{Tabulate an IP's package downloads.}
\usage{
ipPackage(ip = 10, date = NULL, all.filters = FALSE, ip.filter = FALSE,
  triplet.filter = FALSE, small.filter = FALSE, sequence.filter = FALSE,
  size.filter = FALSE, sort = TRUE, memoization = TRUE,
  multi.core = TRUE, dev.mode = FALSE)
}
\arguments{
\item{ip}{Numeric. ip_id.}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{small.filter}{Logical. TRUE filters out downloads less than 1000 bytes.}

\item{sequence.filter}{Logical.}

\item{size.filter}{Logical.}

\item{sort}{Logical. Sort by download count.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
\note{
ip = 10 is a tw top-level domain on 2020-07-09.
}
