% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{localTime}
\alias{localTime}
\title{Compute Local Time from Coordinated Universal Time (UTC/GMT).}
\usage{
localTime(date = "2021-1-1", time = "12:00", tz = Sys.timezone())
}
\arguments{
\item{date}{Character. Date "yyyy-mm-dd".}

\item{time}{Character. Local time "hh:mm" or "hh:mm:ss".}

\item{tz}{Character. Local time zone. See OlsonNames() or use Sys.timezone().}
}
\description{
Compute Local Time from Coordinated Universal Time (UTC/GMT).
}
