% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDate.R
\name{logDate}
\alias{logDate}
\title{Compute Effective CRAN Log Date Based on Local and UTC Time (prototype).}
\usage{
logDate(date = NULL, check.url = TRUE, tz = Sys.timezone(),
  upload.time = "17:00", warning.msg = TRUE, fix.date = TRUE)
}
\arguments{
\item{date}{Character. Date of desired log \code{"yyyy-mm-dd"}. NULL returns date of latest available log.}

\item{check.url}{Logical.}

\item{tz}{Character. Time zone. See OlsonNames().}

\item{upload.time}{Character. UTC upload time for logs "hh:mm" or "hh:mm:ss".}

\item{warning.msg}{Logical. TRUE uses warning() if the function returns the date of the previous available log.}

\item{fix.date}{Logical. Fix date when directly accessing RStudio logs.}
}
\value{
An R date object.
}
\description{
RStudio CRAN Mirror Logs for previous day are posted at 17:00:00 UTC.
}
