% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{logPostInfo}
\alias{logPostInfo}
\title{Compute Date of Log, GMT and Local Posting Time.}
\usage{
logPostInfo(tz = Sys.timezone())
}
\arguments{
\item{tz}{Character. Local time zone. See OlsonNames() or use Sys.timezone().}
}
\description{
Compute Date of Log, GMT and Local Posting Time.
}
