% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{logPostInfo}
\alias{logPostInfo}
\title{Compute Date and Time of Latest Available Log.}
\usage{
logPostInfo(tz = Sys.timezone())
}
\arguments{
\item{tz}{Character. Local time zone. See OlsonNames() or use Sys.timezone().}
}
\description{
GMT and Local Posting Times.
}
