% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageArchive}
\alias{packageArchive}
\title{Extract version history from Archive.}
\usage{
packageArchive(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
An R data frame or NULL.
}
\description{
Extract version history from Archive.
}
\examples{
\donttest{
packageArchive(package = "HistData")
packageArchive(package = "adjustedcranlogs")  # No archived versions.
}
}
