% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageArchive}
\alias{packageArchive}
\title{Scrape package data from Archive.}
\usage{
packageArchive(package = "cholera", check.package = TRUE, size = FALSE)
}
\arguments{
\item{package}{Character. Package name.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{size}{Logical. Include size of source file.}
}
\value{
An R data frame or NULL.
}
\description{
Scrape package data from Archive.
}
\examples{
\dontrun{
packageArchive(package = "HistData")
packageArchive(package = "adjustedcranlogs")  # No archived versions.
}
}
