% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCountry.R
\name{packageCountry}
\alias{packageCountry}
\title{Package download counts by country.}
\usage{
packageCountry(packages = NULL, date = Sys.Date() - 1,
  memoization = TRUE, sort = TRUE, na.rm = FALSE, small.filter = TRUE,
  triplet.filter = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. yyyy-mm-dd.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{sort}{Logical. Sort by download count.}

\item{na.rm}{Logical. Remove NAs.}

\item{small.filter}{Logical.}

\item{triplet.filter}{Logical.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
