% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCountry.R
\name{packageCountry}
\alias{packageCountry}
\title{Package download counts by country.}
\usage{
packageCountry(packages = "cholera", date = NULL, all.filters = FALSE,
  ip.filter = FALSE, triplet.filter = FALSE, small.filter = FALSE,
  sequence.filter = FALSE, size.filter = FALSE, sort = TRUE,
  na.rm = FALSE, memoization = TRUE, check.package = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{all.filters}{Logical. Master switch for filters.}

\item{ip.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{size.filter}{Logical.}

\item{sort}{Logical. Sort by download count.}

\item{na.rm}{Logical. Remove NAs.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
