% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageCRAN}
\alias{packageCRAN}
\title{Scrape package data from CRAN.}
\usage{
packageCRAN(package = "cholera", check.package = TRUE, size = FALSE)
}
\arguments{
\item{package}{Character. Package name.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{size}{Logical. Include size of source file.}
}
\value{
An R data frame or NULL.
}
\description{
Version, date and size (source file) of most recent publication.
}
\examples{
\dontrun{
packageCRAN(package = "HistData")
packageCRAN(package = "VR") # No version on CRAN (archived)
}
}
