% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageCRAN}
\alias{packageCRAN}
\title{Extract package version history from CRAN.}
\usage{
packageCRAN(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
An R data frame or NULL.
}
\description{
Date and version of most recent publication.
}
\examples{
\donttest{
packageCRAN(package = "HistData")
packageCRAN(package = "VR") # No version on CRAN (archived)
}
}
