% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDistribution.R
\name{packageDistribution}
\alias{packageDistribution}
\title{Package Download Distribution Plot.}
\usage{
packageDistribution(package = "HistData", date = Sys.Date() - 1,
  memoization = TRUE, check.cran = FALSE, check.archive = FALSE)
}
\arguments{
\item{package}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd".}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.cran}{Logical. Check if package exists.}

\item{check.archive}{Logical. Include archive when validating package.}
}
\description{
Package Download Distribution Plot.
}
