% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageHistory}
\alias{packageHistory}
\title{Extract package or R version history.}
\usage{
packageHistory(package = "cholera", check.package = FALSE)
}
\arguments{
\item{package}{Character. Vector of package names (including "R").}

\item{check.package}{Logical. Validate and "spell check" package.}
}
\description{
Date and version of all publications.
}
