% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageHistory}
\alias{packageHistory}
\title{Extract package version history CRAN and Archive.}
\usage{
packageHistory(package = "cholera", short.date = TRUE)
}
\arguments{
\item{package}{Character. Package name.}

\item{short.date}{Logical}
}
\description{
Date and version of all publications.
}
