% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHistory.R
\name{packageHistory0}
\alias{packageHistory0}
\title{Extract package version history CRAN and Archive (scrape CRAN).}
\usage{
packageHistory0(package = "cholera")
}
\arguments{
\item{package}{Character. Package name.}
}
\description{
Date and version of most recent publication.
}
