% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\name{packageInfo}
\alias{packageInfo}
\title{Extract packge source file information from CRAN.}
\usage{
packageInfo(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}
}
\description{
Extract packge source file information from CRAN.
}
