% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\name{packageInfo}
\alias{packageInfo}
\title{Extract package information from CRAN.}
\usage{
packageInfo(multi.core = TRUE, platform = "win", r.ver = "release",
  source = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}

\item{platform}{Character.}

\item{r.ver}{Character.}

\item{source}{Logical.}
}
\description{
Extract package information from CRAN.
}
