% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{packageLog}
\alias{packageLog}
\title{Get Package Download Logs.}
\usage{
packageLog(packages = "cholera", date = NULL, ip.filter = TRUE,
  triplet.filter = TRUE, small.filter = TRUE, sequence.filter = TRUE,
  size.filter = TRUE, memoization = TRUE, check.package = TRUE,
  clean.output = FALSE, multi.core = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd". NULL uses latest available log.}

\item{ip.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{size.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{clean.output}{Logical. NULL row names.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
