% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{packageLog}
\alias{packageLog}
\title{Get Package Download Logs.}
\usage{
packageLog(packages = NULL, date = Sys.Date() - 1, small.filter = TRUE,
  triplet.filter = TRUE, ip.filter = TRUE, memoization = TRUE,
  check.package = TRUE, dev.mode = FALSE, clean.output = FALSE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{small.filter}{Logical.}

\item{triplet.filter}{Logical.}

\item{ip.filter}{Logical.}

\item{memoization}{Logical. Use memoization when downloading logs.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{dev.mode}{Logical. Use validatePackage0() to scrape CRAN.}

\item{clean.output}{Logical. NULL row names.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
