% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{packageLog}
\alias{packageLog}
\title{Get Package Download Logs.}
\usage{
packageLog(packages = NULL, date = Sys.Date() - 1, filter = FALSE,
  memoization = TRUE)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}

\item{date}{Character. Date.}

\item{filter}{Logical or Numeric. If Logical, TRUE filters out downloads less than 1000 bytes. If Numeric, a positive value (bytes) sets the minimum download size to consider; a negative value sets the maximum download size to consider.}

\item{memoization}{Logical. Use memoization when downloading logs.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
