% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog2.R
\name{packageLog0}
\alias{packageLog0}
\title{Get Package Download Logs.}
\usage{
packageLog0(date = Sys.Date() - 1)
}
\arguments{
\item{date}{Character. Date.}
}
\value{
An R data frame.
}
\description{
From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
