% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageMRAN.R
\name{packageMRAN}
\alias{packageMRAN}
\title{Extract package data from MRAN (prototype).}
\usage{
packageMRAN(package = "cholera", date = NULL, check.package = TRUE,
  multi.core = TRUE)
}
\arguments{
\item{package}{Character. Package name.}

\item{date}{Character. NULL uses latest available log.}

\item{check.package}{Logical. Validate and "spell check" package.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
Binary or source size.
}
\note{
Depending on when synchronization occurred, you may need to add 3 or 4 days to CRAN publication date, see packageHistory(), to find the package or version you're looking for.
}
