% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank-package.R
\docType{package}
\name{packageRank-package}
\alias{packageRank-package}
\title{packageRank}
\description{
Computation and Visualization of Package Download Counts and Percentiles.
}
\details{
\itemize{
\item Download dounts via cranDownloads().
\item Download rank percentiles via packageRank().
\item Inflation filters via function arguments.
\item Availability of results via logInfo().
\item Data Fixes: early logs and later instance of doubling/tripling.
\item et cetera: country code top-level domains, memoization and download spikes.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lindbrook/packageRank}
  \item Report bugs at \url{https://github.com/lindbrook/packageRank/issues}
}

}
\author{
\strong{Maintainer}: lindbrook \email{lindbrook@gmail.com}

}
