% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank-package.R
\docType{package}
\name{packageRank-package}
\alias{packageRank-package}
\title{packageRank}
\description{
Compute and Visualize Package Download Counts and Percentile Ranks.
}
\details{
\itemize{
\item Download counts via cranDownloads().
\item Percentiles ranks  of download counts via packageRank() and packageLog().
\item Download count inflation filters.
\item Availability of results via logInfo().
\item Reverse lookup of counts, ranks and percentile ranks.
\item Data fixes for early logs and a later instance of double/triple counting of R application downloads.
\item Note Sunday and Wednesday spikes in Windows R Application Nov 2022 - Mar 2023.
\item Country code top-level domains, memoization and internet connection timeout problem.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lindbrook/packageRank}
  \item Report bugs at \url{https://github.com/lindbrook/packageRank/issues}
}

}
\author{
\strong{Maintainer}: lindbrook \email{lindbrook@gmail.com}

}
