% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank-package.R
\docType{package}
\name{packageRank-package}
\alias{packageRank-package}
\title{packageRank}
\description{
Compute and Visualize Package Download Counts and Percentile Ranks.
}
\details{
\itemize{
\item Download counts via cranDownloads().
\item Percentile ranks of download counts via packageRank() and packageLog().
\item Download count inflation filters.
\item Availability of results via logInfo().
\item Reverse lookup of counts, ranks and percentile ranks.
\item Data fixes for logs and 'cranlogs' R package.
\item Note Sunday and Wednesday spikes in Windows R Application Nov 2022 - Mar 2023.
\item Discuss country code top-level domains, memoization and internet connection timeout problem.
\item Graphical fix for missing logs: August 25-26, 29-31 2025; September 1-2, 2025.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lindbrook/packageRank}
  \item Report bugs at \url{https://github.com/lindbrook/packageRank/issues}
}

}
\author{
\strong{Maintainer}: lindbrook \email{lindbrook@gmail.com}

}
