% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank-package.R
\docType{package}
\name{packageRank-package}
\alias{packageRank-package}
\title{packageRank}
\description{
Compute and Visualize Package Download Counts and Percentiles.
}
\details{
\itemize{
\item Download counts via cranDownloads().
\item Rank percentiles of download counts via packageRank() and packageLog().
\item Download count inflation filters.
\item Availability of results via logInfo().
\item Data fixes for early logs and a later instance of double/triple counting of R application downloads.
\item Note Sunday and Wednesday spikes in Windows R Application Nov 2022 - Mar 2023.
\item Country code top-level domains, memoization and internet connection timeout problem.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lindbrook/packageRank}
  \item Report bugs at \url{https://github.com/lindbrook/packageRank/issues}
}

}
\author{
\strong{Maintainer}: lindbrook \email{lindbrook@gmail.com}

}
