% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{packageRankTime}
\alias{packageRankTime}
\title{Package download counts and rank percentiles (longitudinal).}
\usage{
packageRankTime(package = "HistData", when = "last-month",
  sample.pct = 5, multi.core = FALSE)
}
\arguments{
\item{package}{Character. Character. Vector of package name(s).}

\item{when}{Character. "last-month" or "last-week".}

\item{sample.pct}{Numeric.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. On Windows, only \code{multi.core = FALSE} is available.}
}
\description{
Temporal pattern over last week or month.
}
