% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVersionPercent.R
\name{packageVersionPercent}
\alias{packageVersionPercent}
\title{Compute data for versionPlot().}
\usage{
packageVersionPercent(lst, yr.mo = "2020-07", multi.core = TRUE)
}
\arguments{
\item{lst}{Object. List of CRAN download logs data frames. Use monthlyLog().}

\item{yr.mo}{Character. "yyyy-mo". packageVersionsPercent(NULL, yr.mo)}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}
}
\description{
packageRank::blog.data or recompute random sample of packages.
}
\examples{
\dontrun{
# To resample and recompute, set lst to NULL, specify a yr.mo:
packageVersionPercent(NULL, yr.mo = "2020-07")

Otherwise, you must provide a pre-computed lst of logs.
}
}
