% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pkgLog}
\alias{pkgLog}
\title{Extract Package Logs.}
\usage{
pkgLog(lst, i = 1, triplet.filter = TRUE, ip.filter = TRUE,
  ip.campaigns = TRUE, small.filter = TRUE, sequence.filter = TRUE,
  pkg = "cholera", multi.core = TRUE, clean.output = TRUE)
}
\arguments{
\item{lst}{Object. List of logs.}

\item{i}{Numeric. Day/ID.}

\item{triplet.filter}{Logical.}

\item{ip.filter}{Logical.}

\item{ip.campaigns}{Logical.}

\item{small.filter}{Logical.}

\item{sequence.filter}{Logical.}

\item{pkg}{Character.}

\item{multi.core}{Logical or Numeric.}

\item{clean.output}{Logical.}
}
\description{
Extract Package Logs.
}
