% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pkgLog0}
\alias{pkgLog0}
\title{Extract Package Logs.}
\usage{
pkgLog0(lst, i = 1, pkg = "cholera", clean.output = TRUE)
}
\arguments{
\item{lst}{Object. List of logs.}

\item{i}{Numeric. Day/ID.}

\item{pkg}{Character.}

\item{clean.output}{Logical.}
}
\description{
Extract Package Logs.
}
