% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualDownloads.R
\name{plot.annualDownloads}
\alias{plot.annualDownloads}
\title{Plot method for annualDownloads().}
\usage{
\method{plot}{annualDownloads}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for annualDownloads().
}
