% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualDownloads.R
\name{plot.annualDownloads}
\alias{plot.annualDownloads}
\title{Plot method for annualDownloads().}
\usage{
\method{plot}{annualDownloads}(x, statistic = "count", pool.obs = FALSE,
  log.y = FALSE, sep.y = FALSE, nrow = 3, smooth = TRUE, span = 3/4,
  ...)
}
\arguments{
\item{x}{object.}

\item{statistic}{Character. "count" or "percent".}

\item{pool.obs}{Logical.}

\item{log.y}{Logical. Base 10 logarithm of y-axis.}

\item{sep.y}{Logical. Separate, independent y-scales for each panel.}

\item{nrow}{Numeric. Number of rows for ggplot2 facets.}

\item{smooth}{Logical. Add smoother (loess).}

\item{span}{Numeric. Smoothing parameter for geom_smooth(); c.f. stats::loess(span). 3/4 is built-in default.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for annualDownloads().
}
