% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualDownloads.R
\name{plot.annualDownloads}
\alias{plot.annualDownloads}
\title{Plot method for annualDownloads().}
\usage{
\method{plot}{annualDownloads}(x, statistic = "count", pool = TRUE,
  log.y = FALSE, sep.y = FALSE, nrow = 3, smooth = TRUE, f = 1/4,
  span = 3/4, points = FALSE, line.col = "gray", ...)
}
\arguments{
\item{x}{object.}

\item{statistic}{Character. "count" or "percent".}

\item{pool}{Logical. Pool annual data into single time series.}

\item{log.y}{Logical. Base 10 logarithm of y-axis.}

\item{sep.y}{Logical. Separate, independent y-scales for each panel.}

\item{nrow}{Numeric. Number of rows for ggplot2 facets.}

\item{smooth}{Logical. Add smoother (loess).}

\item{f}{Numeric. Parameter for lowess.}

\item{span}{Numeric. Smoothing parameter for geom_smooth(), which uses loess.}

\item{points}{Logical.}

\item{line.col}{Character. Color of line}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for annualDownloads().
}
