% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads.R
\name{plot.bioconductor}
\alias{plot.bioconductor}
\title{Plot method for cran_downloads2().}
\usage{
\method{plot}{bioconductor}(x, count = "download", smooth = TRUE, ...)
}
\arguments{
\item{x}{object.}

\item{count}{Character. "download" or "ip".}

\item{smooth}{Logical. Add stats::lowess smoother.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for cran_downloads2().
}
