% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductorDownloads.R
\name{plot.bioconductor}
\alias{plot.bioconductor}
\title{Plot method for bioconductor_downloads().}
\usage{
\method{plot}{bioconductor}(x, graphics = NULL, count = "download",
  add.points = TRUE, smooth = FALSE, smooth.f = 2/3, se = FALSE,
  log_count = FALSE, ...)
}
\arguments{
\item{x}{object.}

\item{graphics}{Character. NULL, "base" or "ggplot2".}

\item{count}{Character. "download" or "ip".}

\item{add.points}{Logical. Add points.}

\item{smooth}{Logical. Add stats::lowess smoother.}

\item{smooth.f}{Numeric. smoother span.}

\item{se}{Logical. Works only with graphics = "ggplot2".}

\item{log_count}{Logical. Logarithm of package downloads.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for bioconductor_downloads().
}
\examples{
\donttest{
plot(bioconductorDownloads())
plot(bioconductorDownloads(packages = "graph"))
plot(bioconductorDownloads(packages = "graph", year = 2019))
plot(bioconductorDownloads(packages = "graph", year = 2014, end.year = 2018, month = 6, end.month = 3))
}
}
