% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_downloads2.R
\name{plot.bioconductor2}
\alias{plot.bioconductor2}
\title{Plot method for bioconductor_downloads().}
\usage{
\method{plot}{bioconductor2}(x, count = "download", smooth = TRUE, ...)
}
\arguments{
\item{x}{object.}

\item{count}{Character. "download" or "ip".}

\item{smooth}{Logical. Add stats::lowess smoother.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for bioconductor_downloads().
}
