% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDistribution.R
\name{plot.cranDistribution}
\alias{plot.cranDistribution}
\title{Plot method for cranDistribution().}
\usage{
\method{plot}{cranDistribution}(x, type = "count", ...)
}
\arguments{
\item{x}{An object of class "cranDistribution" created by \code{cranDistribution()}.}

\item{type}{Character. "histogram" or "count".}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for cranDistribution().
}
