% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRankTime.R
\name{plot.package_rank_time}
\alias{plot.package_rank_time}
\title{Plot method for timeSeriesRank().}
\usage{
\method{plot}{package_rank_time}(x, graphics_pkg = "ggplot2",
  log_y = TRUE, cran_smpl_smooth = TRUE, f = 1/3, ...)
}
\arguments{
\item{x}{Object. An object of class "time_series" created by \code{packageRankTime()}.}

\item{graphics_pkg}{Character. "base" or "ggplot2".}

\item{log_y}{Logical.}

\item{cran_smpl_smooth}{Logical. lowess background.}

\item{f}{Numeric. lowess smoothing window}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for timeSeriesRank().
}
