% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{plot.packageLog}
\alias{plot.packageLog}
\title{Plot method for packageLog().}
\usage{
\method{plot}{packageLog}(x, type = "1D", unit.observation = "second", ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "1D" or "2D".}

\item{unit.observation}{Character. "second", "minute", or "hour".}

\item{...}{Additional parameters.}
}
\description{
Plot method for packageLog().
}
