% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLog.R
\name{plot.packageLog}
\alias{plot.packageLog}
\title{Plot method for packageLog().}
\usage{
\method{plot}{packageLog}(x, type = "1D", time.unit = "second",
  smooth = FALSE, points = TRUE, same.xy = TRUE,
  local.timezone = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "1D" or "2D".}

\item{time.unit}{Character. "second", "minute", or "hour".}

\item{smooth}{Logical. Add smoother.}

\item{points}{Logical. For "hour" and "minute" in 2D plots.}

\item{same.xy}{Logical. Use same scale for multiple packages for type = "2D".}

\item{local.timezone}{Logical or Character. TRUE for Sys.timezone(). See OlsonNames() for other time zones.}

\item{...}{Additional parameters.}
}
\description{
Plot method for packageLog().
}
