% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRank.R
\name{plot.packageRank}
\alias{plot.packageRank}
\title{Plot method for packageRank() and packageRank0().}
\usage{
\method{plot}{packageRank}(x, graphics = NULL, log.y = TRUE, ...)
}
\arguments{
\item{x}{An object of class "packageRank" created by \code{packageRank()}.}

\item{graphics}{Character. "base" or "ggplot2".}

\item{log.y}{Logical. Logarithm of package downloads.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R or ggplot2 plot.
}
\description{
Plot method for packageRank() and packageRank0().
}
\examples{
\dontrun{
plot(packageRank(packages = "HistData", date = "2020-01-01"))
plot(packageRank(packages = c("h2o", "Rcpp", "rstan"), date = "2020-01-01"))
}
}
