% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyDownloads.R
\name{plot.weeklyDownloads}
\alias{plot.weeklyDownloads}
\title{Plot method for annualDownloads().}
\usage{
\method{plot}{weeklyDownloads}(x, statistic = "percent",
  aggregation = "day", typical.value = "mean", nrow = 3L, ...)
}
\arguments{
\item{x}{object.}

\item{statistic}{Character. "count" or "percent".}

\item{aggregation}{Character. "week" or "day".}

\item{typical.value}{Character. "mean" or "median".}

\item{nrow}{Numeric. Number of rows for ggplot2 facets.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for annualDownloads().
}
\examples{
\dontrun{
plot(weeklyDownloads())
plot(weeklyDownloads(n = 9), aggregation = "week")
}
}
