% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCountry.R
\name{plotTopCountryCodes}
\alias{plotTopCountryCodes}
\title{Plot Top N Downloads by Country Code.}
\usage{
plotTopCountryCodes(dat = packageRank::blog.data$top.n)
}
\arguments{
\item{dat}{Object.}
}
\description{
Plot Top N Downloads by Country Code.
}
