% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCountry.R
\name{plotTopCountryCodes}
\alias{plotTopCountryCodes}
\title{Plot Top N Downloads by Country Code.}
\usage{
plotTopCountryCodes(dataset = "october", second.place = FALSE)
}
\arguments{
\item{dataset}{Character.}

\item{second.place}{Logical. Annotate second place country.}
}
\description{
Plot Top N Downloads by Country Code.
}
