% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranDistribution.R
\name{print.cranDistribution}
\alias{print.cranDistribution}
\title{Print method for cranDistribution().}
\usage{
\method{print}{cranDistribution}(x, top.n = 20, ...)
}
\arguments{
\item{x}{object.}

\item{top.n}{Numeric or Integer.}

\item{...}{Additional parameters.}
}
\description{
Print method for cranDistribution().
}
