% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveDate.R
\name{resolveDate}
\alias{resolveDate}
\title{Resolve date.}
\usage{
resolveDate(date, type = "from")
}
\arguments{
\item{date}{Character. \code{"yyyy-mm-dd"}, \code{"yyyy-mm"}, \code{"yyyy"} or \code{yyyy} (numeric).}

\item{type}{Character. Type of date "to" or "from".}
}
\description{
Check date format and validate date.
}
