% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFilter.R
\name{sequenceFilter}
\alias{sequenceFilter}
\title{Filter Downloads of sequential Versions (prototype).}
\usage{
sequenceFilter(dat, timeframe = 15, one.timeframe = TRUE,
  one.country = TRUE, one.ip = TRUE)
}
\arguments{
\item{dat}{Object.}

\item{timeframe}{Numeric. minutes}

\item{one.timeframe}{Logical.}

\item{one.country}{Logical.}

\item{one.ip}{Logical.}
}
\description{
Filter Downloads of sequential Versions (prototype).
}
