% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFilter.R
\name{sequenceFilter}
\alias{sequenceFilter}
\title{Filter downloads of full-sized sequential versions (prototype).}
\usage{
sequenceFilter(pkg.data, arch.pkg.history, download.time = 30)
}
\arguments{
\item{pkg.data}{Object.}

\item{arch.pkg.history}{Object.}

\item{download.time}{Numeric. Package download time allowance (seconds).}
}
\description{
Filter downloads of full-sized sequential versions (prototype).
}
