% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFilter.R
\name{sequenceFilter}
\alias{sequenceFilter}
\title{Filter downloads of full-sized sequential versions (prototype).}
\usage{
sequenceFilter(dat, packages, ymd, cores, download.time = 30,
  dev.mode = dev.mode)
}
\arguments{
\item{dat}{Object.}

\item{packages}{Object. An R vector of package names.}

\item{ymd}{Date. Log date.}

\item{cores}{Numeric. Number of cores to use.}

\item{download.time}{Numeric. Package download time allowance (seconds).}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
Filter downloads of full-sized sequential versions (prototype).
}
