% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialVersion.R
\name{sequentialVersion}
\alias{sequentialVersion}
\title{Download of sequential versions}
\usage{
sequentialVersion(package = "cholera", date = Sys.Date() - 1,
  time.frame = 60, one.country = FALSE, one.ip = FALSE)
}
\arguments{
\item{package}{Character. Vector of package name(s).}

\item{date}{Character. Date. "yyyy-mm-dd".}

\item{time.frame}{Numeric. seconds.}

\item{one.country}{Logical.}

\item{one.ip}{Logical.}
}
\description{
Download of sequential versions
}
