% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallFilter.R
\name{smallFilter}
\alias{smallFilter}
\title{Filter out small downloads (prototype).}
\usage{
smallFilter(dat, threshold = 1000L, multi.core = TRUE,
  dev.mode = dev.mode)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{threshold}{Numeric. Bytes.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
Filter out small downloads (prototype).
}
