% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallFilter.R
\name{smallFilter}
\alias{smallFilter}
\title{Filter out small downloads (prototype).}
\usage{
smallFilter(dat, threshold = 1000L)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{threshold}{Numeric. Bytes.}
}
\description{
Filter out small downloads (prototype).
}
