% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripletFilter.R
\name{tripletFilter}
\alias{tripletFilter}
\title{Filter out small downloads triplets (prototype).}
\usage{
tripletFilter(dat, time.window = 2)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{time.window}{Numeric. Seconds.}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
