% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripletFilter.R
\name{tripletFilter}
\alias{tripletFilter}
\title{Filter out small downloads triplets (prototype).}
\usage{
tripletFilter(dat, time.window = 2, multi.core = TRUE,
  dev.mode = dev.mode)
}
\arguments{
\item{dat}{Object. Package log entries.}

\item{time.window}{Numeric. Seconds.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
Logs from RStudio's CRAN Mirror http://cran-logs.rstudio.com/
}
