% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePackage.R
\name{validatePackage}
\alias{validatePackage}
\title{Check for valid package names.}
\usage{
validatePackage(packages)
}
\arguments{
\item{packages}{Character. Vector of package name(s).}
}
\description{
Check for valid package names.
}
